Refactor instructions
