+++ programs/token/src/instructions/burn.rs
@@ -1,10 +1,11 @@
 use {
     crate::{
-        definitions::{
-            account_borrow_failed_error, invalid_argument_error, CpiWriter, TokenProgram,
-            MAX_MULTISIG_SIGNERS,
+        instructions::{
+            account_borrow_failed_error, initialize_multisig::MAX_MULTISIG_SIGNERS,
+            invalid_argument_error, write_bytes, CpiWriter, UNINIT_BYTE, UNINIT_CPI_ACCOUNT,
+            UNINIT_INSTRUCTION_ACCOUNT,
         },
-        write_bytes, UNINIT_BYTE, UNINIT_CPI_ACCOUNT, UNINIT_INSTRUCTION_ACCOUNT,
+        TokenProgram,
     },
     core::{marker::PhantomData, mem::MaybeUninit, slice::from_raw_parts},
     solana_account_view::AccountView,
@@ -17,112 +18,121 @@ use {
 };
 
 /// The instruction discriminator.
-const DISCRIMINATOR: u8 = 6;
+const DISCRIMINATOR: u8 = 8;
 
 /// Maximum number of accounts expected by this instruction.
 ///
 /// The required number of accounts will depend whether the
 /// source account has a single owner or a multisignature
 /// owner.
-const MAX_ACCOUNTS_LEN: usize = 2 + MAX_MULTISIG_SIGNERS;
+const MAX_ACCOUNTS_LEN: usize = 3 + MAX_MULTISIG_SIGNERS;
 
 /// Instruction data length:
 ///   - discriminator (1 byte)
-///   - authority type (1 byte)
-///   - new authority (33 bytes, optional)
-const DATA_LEN: usize = 35;
+///   - amount (8 bytes)
+const DATA_LEN: usize = 9;
 
-#[repr(u8)]
-#[derive(Clone, Copy)]
-pub enum AuthorityType {
-    MintTokens = 0,
-    FreezeAccount = 1,
-    AccountOwner = 2,
-    CloseAccount = 3,
-}
-
-/// Sets a new authority of a mint or account.
+/// Burns tokens by removing them from an account.  `Burn` does not support
+/// accounts associated with the native mint, use `CloseAccount` instead.
 ///
 /// Accounts expected by this instruction:
 ///
-///   * Single authority
-///   0. `[writable]` The mint or account to change the authority of.
-///   1. `[signer]` The current authority of the mint or account.
+///   * Single owner/delegate
+///   0. `[writable]` The account to burn from.
+///   1. `[writable]` The token mint.
+///   2. `[signer]` The account's owner/delegate.
 ///
-///   * Multisignature authority
-///   0. `[writable]` The mint or account to change the authority of.
-///   1. `[]` The mint's or account's current multisignature authority.
-///   2. `..+M` `[signer]` M signer accounts.
-pub struct SetAuthority<
-    'account,
-    'address,
-    'multisig,
-    MultisigSigner: AsRef<AccountView>,
-    Program: TokenProgram,
-> {
-    /// The mint or account to change the authority of.
+///   * Multisignature owner/delegate
+///   0. `[writable]` The account to burn from.
+///   1. `[writable]` The token mint.
+///   2. `[]` The account's multisignature owner/delegate.
+///   3. `..+M` `[signer]` M signer accounts.
+pub struct Burn<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram> {
+    /// The account to burn from.
     pub account: &'account AccountView,
 
-    ///  The current authority of the mint or account.
+    /// The token mint.
+    pub mint: &'account AccountView,
+
+    /// The account's owner/delegate.
     pub authority: &'account AccountView,
 
     /// Multisignature signers.
     pub multisig_signers: &'multisig [MultisigSigner],
 
-    /// The type of authority to update.
-    pub authority_type: AuthorityType,
-
-    /// The new authority.
-    pub new_authority: Option<&'address Address>,
+    /// The amount of tokens to burn.
+    pub amount: u64,
 
     _program: PhantomData<Program>,
 }
 
-impl<'account, 'address, Program: TokenProgram>
-    SetAuthority<'account, 'address, '_, &'account AccountView, Program>
-{
-    /// Creates a new `SetAuthority` instruction with a single authority.
+impl<'account, Program: TokenProgram> Burn<'account, '_, &'account AccountView, Program> {
+    /// The instruction discriminator.
+    pub const DISCRIMINATOR: u8 = DISCRIMINATOR;
+
+    /// Maximum number of accounts expected by this instruction.
+    pub const MAX_ACCOUNTS_LEN: usize = MAX_ACCOUNTS_LEN;
+
+    /// Instruction data length.
+    pub const DATA_LEN: usize = DATA_LEN;
+
+    /// Creates a new `Burn` instruction with a single
+    /// owner/delegate authority.
     #[inline(always)]
     pub fn new(
         account: &'account AccountView,
+        mint: &'account AccountView,
         authority: &'account AccountView,
-        authority_type: AuthorityType,
-        new_authority: Option<&'address Address>,
+        amount: u64,
     ) -> Self {
-        Self::with_multisig_signers(account, authority, authority_type, new_authority, &[])
+        Self::with_multisig_signers(account, mint, authority, amount, &[])
     }
 }
 
-impl<'account, 'address, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram>
-    SetAuthority<'account, 'address, 'multisig, MultisigSigner, Program>
+impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram>
+    Burn<'account, 'multisig, MultisigSigner, Program>
 {
-    /// Creates a new `SetAuthority` instruction with a
-    /// multisignature authority and signer accounts.
+    /// Creates a new `Burn` instruction with a
+    /// multisignature owner/delegate authority and signer accounts.
     #[inline(always)]
     pub fn with_multisig_signers(
         account: &'account AccountView,
+        mint: &'account AccountView,
         authority: &'account AccountView,
-        authority_type: AuthorityType,
-        new_authority: Option<&'address Address>,
+        amount: u64,
         multisig_signers: &'multisig [MultisigSigner],
     ) -> Self {
         Self {
             account,
+            mint,
             authority,
             multisig_signers,
-            authority_type,
-            new_authority,
+            amount,
             _program: PhantomData,
         }
     }
 
     #[inline(always)]
     pub fn invoke(&self) -> ProgramResult {
-        self.invoke_signed(&[])
+        self.invoke_with_program(&Program::ID)
     }
 
     #[inline(always)]
     pub fn invoke_signed(&self, signers: &[Signer]) -> ProgramResult {
+        self.invoke_signed_with_program(signers, &Program::ID)
+    }
+
+    #[inline(always)]
+    pub fn invoke_with_program(&self, program: &Address) -> ProgramResult {
+        self.invoke_signed_with_program(&[], program)
+    }
+
+    #[inline(always)]
+    pub fn invoke_signed_with_program(
+        &self,
+        signers: &[Signer],
+        program: &Address,
+    ) -> ProgramResult {
         if self.multisig_signers.len() > MAX_MULTISIG_SIGNERS {
             Err(ProgramError::InvalidArgument)?;
         }
@@ -140,7 +150,7 @@ impl<'account, 'address, 'multisig, MultisigSigner: AsRef<AccountView>, Program:
         unsafe {
             invoke_signed_unchecked(
                 &InstructionView {
-                    program_id: &Program::id(),
+                    program_id: program,
                     accounts: from_raw_parts(
                         instruction_accounts.as_ptr() as _,
                         written_instruction_accounts,
@@ -157,7 +167,7 @@ impl<'account, 'address, 'multisig, MultisigSigner: AsRef<AccountView>, Program:
 }
 
 impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
-    for SetAuthority<'_, '_, '_, MultisigSigner, Program>
+    for Burn<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn write_accounts<'cpi>(
@@ -169,6 +179,7 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
     {
         write_accounts(
             self.account,
+            self.mint,
             self.authority,
             self.multisig_signers,
             accounts,
@@ -185,6 +196,7 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
     {
         write_instruction_accounts(
             self.account,
+            self.mint,
             self.authority,
             self.multisig_signers,
             accounts,
@@ -193,17 +205,17 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
 
     #[inline(always)]
     fn write_instruction_data(&self, data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
-        write_instruction_data(self.authority_type, self.new_authority, data)
+        write_instruction_data(self.amount, data)
     }
 }
 
-impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch<Program>
-    for SetAuthority<'_, '_, '_, MultisigSigner, Program>
+impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::batch::IntoBatch<Program>
+    for Burn<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn into_batch<'account, 'state>(
         self,
-        batch: &mut super::Batch<'account, 'state, Program>,
+        batch: &mut super::batch::Batch<'account, 'state, Program>,
     ) -> ProgramResult
     where
         Self: 'account + 'state,
@@ -212,6 +224,7 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch
             |accounts| {
                 write_accounts(
                     self.account,
+                    self.mint,
                     self.authority,
                     self.multisig_signers,
                     accounts,
@@ -220,12 +233,13 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch
             |accounts| {
                 write_instruction_accounts(
                     self.account,
+                    self.mint,
                     self.authority,
                     self.multisig_signers,
                     accounts,
                 )
             },
-            |data| write_instruction_data(self.authority_type, self.new_authority, data),
+            |data| write_instruction_data(self.amount, data),
         )
     }
 }
@@ -233,6 +247,7 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch
 #[inline(always)]
 fn write_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
     account: &'account AccountView,
+    mint: &'account AccountView,
     authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<CpiAccount<'out>>],
@@ -241,21 +256,23 @@ where
     'account: 'out,
     'multisig: 'out,
 {
-    let expected_accounts = 2 + multisig_signers.len();
+    let expected_accounts = 3 + multisig_signers.len();
 
     if expected_accounts > accounts.len() {
         return Err(invalid_argument_error());
     }
 
-    if account.is_borrowed() {
+    if account.is_borrowed() | mint.is_borrowed() {
         return Err(account_borrow_failed_error());
     }
 
     CpiAccount::init_from_account_view(account, &mut accounts[0]);
 
-    CpiAccount::init_from_account_view(authority, &mut accounts[1]);
+    CpiAccount::init_from_account_view(mint, &mut accounts[1]);
 
-    for (account, signer) in accounts[2..expected_accounts]
+    CpiAccount::init_from_account_view(authority, &mut accounts[2]);
+
+    for (account, signer) in accounts[3..expected_accounts]
         .iter_mut()
         .zip(multisig_signers.iter())
     {
@@ -268,6 +285,7 @@ where
 #[inline(always)]
 fn write_instruction_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
     account: &'account AccountView,
+    mint: &'account AccountView,
     authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<InstructionAccount<'out>>],
@@ -276,7 +294,7 @@ where
     'account: 'out,
     'multisig: 'out,
 {
-    let expected_accounts = 2 + multisig_signers.len();
+    let expected_accounts = 3 + multisig_signers.len();
 
     if expected_accounts > accounts.len() {
         return Err(invalid_argument_error());
@@ -284,13 +302,15 @@ where
 
     accounts[0].write(InstructionAccount::writable(account.address()));
 
-    accounts[1].write(InstructionAccount::new(
+    accounts[1].write(InstructionAccount::writable(mint.address()));
+
+    accounts[2].write(InstructionAccount::new(
         authority.address(),
         false,
         multisig_signers.is_empty(),
     ));
 
-    for (account, signer) in accounts[2..expected_accounts]
+    for (account, signer) in accounts[3..expected_accounts]
         .iter_mut()
         .zip(multisig_signers.iter())
     {
@@ -304,31 +324,16 @@ where
 
 #[inline(always)]
 fn write_instruction_data(
-    authority_type: AuthorityType,
-    new_authority: Option<&Address>,
+    amount: u64,
     data: &mut [MaybeUninit<u8>],
 ) -> Result<usize, ProgramError> {
-    if data.len() < 3 {
+    if data.len() < DATA_LEN {
         return Err(invalid_argument_error());
     }
 
     data[0].write(DISCRIMINATOR);
 
-    data[1].write(authority_type as u8);
+    write_bytes(&mut data[1..DATA_LEN], &amount.to_le_bytes());
 
-    if let Some(new_authority) = new_authority {
-        if data.len() < DATA_LEN {
-            return Err(invalid_argument_error());
-        }
-
-        data[2].write(1);
-
-        write_bytes(&mut data[3..DATA_LEN], new_authority.as_array());
-
-        Ok(DATA_LEN)
-    } else {
-        data[2].write(0);
-
-        Ok(3)
-    }
+    Ok(DATA_LEN)
 }
