+++ programs/token/src/instructions/burn_checked.rs
@@ -1,13 +1,15 @@
 use {
     crate::{
-        definitions::{
-            account_borrow_failed_error, invalid_argument_error, CpiWriter, TokenProgram,
-            MAX_MULTISIG_SIGNERS,
+        instructions::{
+            account_borrow_failed_error, initialize_multisig::MAX_MULTISIG_SIGNERS,
+            invalid_argument_error, write_bytes, CpiWriter, UNINIT_BYTE, UNINIT_CPI_ACCOUNT,
+            UNINIT_INSTRUCTION_ACCOUNT,
         },
-        write_bytes, UNINIT_BYTE, UNINIT_CPI_ACCOUNT, UNINIT_INSTRUCTION_ACCOUNT,
+        TokenProgram,
     },
     core::{marker::PhantomData, mem::MaybeUninit, slice::from_raw_parts},
     solana_account_view::AccountView,
+    solana_address::Address,
     solana_instruction_view::{
         cpi::{invoke_signed_unchecked, CpiAccount, Signer},
         InstructionAccount, InstructionView,
@@ -16,14 +18,14 @@ use {
 };
 
 /// The instruction discriminator.
-const DISCRIMINATOR: u8 = 13;
+const DISCRIMINATOR: u8 = 15;
 
 /// Maximum number of accounts expected by this instruction.
 ///
 /// The required number of accounts will depend whether the
 /// source account has a single owner or a multisignature
 /// owner.
-const MAX_ACCOUNTS_LEN: usize = 4 + MAX_MULTISIG_SIGNERS;
+const MAX_ACCOUNTS_LEN: usize = 3 + MAX_MULTISIG_SIGNERS;
 
 /// Instruction data length:
 ///   - discriminator (1 byte)
@@ -31,49 +33,45 @@ const MAX_ACCOUNTS_LEN: usize = 4 + MAX_MULTISIG_SIGNERS;
 ///   - decimals (1 byte)
 const DATA_LEN: usize = 10;
 
-/// Approves a delegate.  A delegate is given the authority over tokens on
-/// behalf of the source account's owner.
+/// Burns tokens by removing them from an account.
+/// [`super::BurnChecked`] does not support accounts
+/// associated with the native mint, use `CloseAccount` instead.
 ///
-/// This instruction differs from Approve in that the token mint and
-/// decimals value is checked by the caller.  This may be useful when
-/// creating transactions offline or within a hardware wallet.
+/// This instruction differs from Burn in that the decimals value is checked
+/// by the caller. This may be useful when creating transactions offline or
+/// within a hardware wallet.
 ///
 /// Accounts expected by this instruction:
 ///
-///   * Single owner
-///   0. `[writable]` The source account.
-///   1. `[]` The token mint.
-///   2. `[]` The delegate.
-///   3. `[signer]` The source account owner.
+///   * Single owner/delegate
+///   0. `[writable]` The account to burn from.
+///   1. `[writable]` The token mint.
+///   2. `[signer]` The account's owner/delegate.
 ///
-///   * Multisignature owner
-///   0. `[writable]` The source account.
-///   1. `[]` The token mint.
-///   2. `[]` The delegate.
-///   3. `[]` The source account's multisignature owner.
-///   4. `..+M` `[signer]` M signer accounts.
-pub struct ApproveChecked<
+///   * Multisignature owner/delegate
+///   0. `[writable]` The account to burn from.
+///   1. `[writable]` The token mint.
+///   2. `[]` The account's multisignature owner/delegate.
+///   3. `..+M` `[signer]` M signer accounts.
+pub struct BurnChecked<
     'account,
     'multisig,
     MultisigSigner: AsRef<AccountView>,
     Program: TokenProgram,
 > {
-    /// The source account.
-    pub source: &'account AccountView,
+    /// The account to burn from.
+    pub account: &'account AccountView,
 
     /// The token mint.
     pub mint: &'account AccountView,
 
-    /// The delegate.
-    pub delegate: &'account AccountView,
-
-    /// The source account owner.
+    /// The account's owner/delegate.
     pub authority: &'account AccountView,
 
     /// Multisignature signers.
     pub multisig_signers: &'multisig [MultisigSigner],
 
-    /// The amount of tokens the delegate is approved for.
+    /// The amount of tokens to burn.
     pub amount: u64,
 
     /// Expected number of base 10 digits to the right of the decimal place.
@@ -82,41 +80,47 @@ pub struct ApproveChecked<
     _program: PhantomData<Program>,
 }
 
-impl<'account, Program: TokenProgram> ApproveChecked<'account, '_, &'account AccountView, Program> {
-    /// Creates a new `ApproveChecked` instruction with a single owner
-    /// authority.
+impl<'account, Program: TokenProgram> BurnChecked<'account, '_, &'account AccountView, Program> {
+    /// The instruction discriminator.
+    pub const DISCRIMINATOR: u8 = DISCRIMINATOR;
+
+    /// Maximum number of accounts expected by this instruction.
+    pub const MAX_ACCOUNTS_LEN: usize = MAX_ACCOUNTS_LEN;
+
+    /// Instruction data length.
+    pub const DATA_LEN: usize = DATA_LEN;
+
+    /// Creates a new `BurnChecked` instruction with a single
+    /// owner/delegate authority.
     #[inline(always)]
     pub fn new(
-        source: &'account AccountView,
+        account: &'account AccountView,
         mint: &'account AccountView,
-        delegate: &'account AccountView,
         authority: &'account AccountView,
         amount: u64,
         decimals: u8,
     ) -> Self {
-        Self::with_multisig_signers(source, mint, delegate, authority, amount, decimals, &[])
+        Self::with_multisig_signers(account, mint, authority, amount, decimals, &[])
     }
 }
 
 impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram>
-    ApproveChecked<'account, 'multisig, MultisigSigner, Program>
+    BurnChecked<'account, 'multisig, MultisigSigner, Program>
 {
-    /// Creates a new `ApproveChecked` instruction with a
-    /// multisignature owner authority and signer accounts.
+    /// Creates a new `BurnChecked` instruction with a
+    /// multisignature owner/delegate authority and signer accounts.
     #[inline(always)]
     pub fn with_multisig_signers(
-        source: &'account AccountView,
+        account: &'account AccountView,
         mint: &'account AccountView,
-        delegate: &'account AccountView,
         authority: &'account AccountView,
         amount: u64,
         decimals: u8,
         multisig_signers: &'multisig [MultisigSigner],
     ) -> Self {
         Self {
-            source,
+            account,
             mint,
-            delegate,
             authority,
             multisig_signers,
             amount,
@@ -127,11 +131,25 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
 
     #[inline(always)]
     pub fn invoke(&self) -> ProgramResult {
-        self.invoke_signed(&[])
+        self.invoke_with_program(&Program::ID)
     }
 
     #[inline(always)]
     pub fn invoke_signed(&self, signers: &[Signer]) -> ProgramResult {
+        self.invoke_signed_with_program(signers, &Program::ID)
+    }
+
+    #[inline(always)]
+    pub fn invoke_with_program(&self, program: &Address) -> ProgramResult {
+        self.invoke_signed_with_program(&[], program)
+    }
+
+    #[inline(always)]
+    pub fn invoke_signed_with_program(
+        &self,
+        signers: &[Signer],
+        program: &Address,
+    ) -> ProgramResult {
         if self.multisig_signers.len() > MAX_MULTISIG_SIGNERS {
             Err(ProgramError::InvalidArgument)?;
         }
@@ -149,7 +167,7 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
         unsafe {
             invoke_signed_unchecked(
                 &InstructionView {
-                    program_id: &Program::id(),
+                    program_id: program,
                     accounts: from_raw_parts(
                         instruction_accounts.as_ptr() as _,
                         written_instruction_accounts,
@@ -166,7 +184,7 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
 }
 
 impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
-    for ApproveChecked<'_, '_, MultisigSigner, Program>
+    for BurnChecked<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn write_accounts<'cpi>(
@@ -177,9 +195,8 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
         Self: 'cpi,
     {
         write_accounts(
-            self.source,
+            self.account,
             self.mint,
-            self.delegate,
             self.authority,
             self.multisig_signers,
             accounts,
@@ -195,9 +212,8 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
         Self: 'cpi,
     {
         write_instruction_accounts(
-            self.source,
+            self.account,
             self.mint,
-            self.delegate,
             self.authority,
             self.multisig_signers,
             accounts,
@@ -210,13 +226,13 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
     }
 }
 
-impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch<Program>
-    for ApproveChecked<'_, '_, MultisigSigner, Program>
+impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::batch::IntoBatch<Program>
+    for BurnChecked<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn into_batch<'account, 'state>(
         self,
-        batch: &mut super::Batch<'account, 'state, Program>,
+        batch: &mut super::batch::Batch<'account, 'state, Program>,
     ) -> ProgramResult
     where
         Self: 'account + 'state,
@@ -224,9 +240,8 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch
         batch.push(
             |accounts| {
                 write_accounts(
-                    self.source,
+                    self.account,
                     self.mint,
-                    self.delegate,
                     self.authority,
                     self.multisig_signers,
                     accounts,
@@ -234,9 +249,8 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch
             },
             |accounts| {
                 write_instruction_accounts(
-                    self.source,
+                    self.account,
                     self.mint,
-                    self.delegate,
                     self.authority,
                     self.multisig_signers,
                     accounts,
@@ -249,9 +263,8 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch
 
 #[inline(always)]
 fn write_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
-    source: &'account AccountView,
+    account: &'account AccountView,
     mint: &'account AccountView,
-    delegate: &'account AccountView,
     authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<CpiAccount<'out>>],
@@ -260,25 +273,23 @@ where
     'account: 'out,
     'multisig: 'out,
 {
-    let expected_accounts = 4 + multisig_signers.len();
+    let expected_accounts = 3 + multisig_signers.len();
 
     if expected_accounts > accounts.len() {
         return Err(invalid_argument_error());
     }
 
-    if source.is_borrowed() {
+    if account.is_borrowed() | mint.is_borrowed() {
         return Err(account_borrow_failed_error());
     }
 
-    CpiAccount::init_from_account_view(source, &mut accounts[0]);
+    CpiAccount::init_from_account_view(account, &mut accounts[0]);
 
     CpiAccount::init_from_account_view(mint, &mut accounts[1]);
 
-    CpiAccount::init_from_account_view(delegate, &mut accounts[2]);
+    CpiAccount::init_from_account_view(authority, &mut accounts[2]);
 
-    CpiAccount::init_from_account_view(authority, &mut accounts[3]);
-
-    for (account, signer) in accounts[4..expected_accounts]
+    for (account, signer) in accounts[3..expected_accounts]
         .iter_mut()
         .zip(multisig_signers.iter())
     {
@@ -290,9 +301,8 @@ where
 
 #[inline(always)]
 fn write_instruction_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
-    source: &'account AccountView,
+    account: &'account AccountView,
     mint: &'account AccountView,
-    delegate: &'account AccountView,
     authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<InstructionAccount<'out>>],
@@ -301,25 +311,23 @@ where
     'account: 'out,
     'multisig: 'out,
 {
-    let expected_accounts = 4 + multisig_signers.len();
+    let expected_accounts = 3 + multisig_signers.len();
 
     if expected_accounts > accounts.len() {
         return Err(invalid_argument_error());
     }
 
-    accounts[0].write(InstructionAccount::writable(source.address()));
+    accounts[0].write(InstructionAccount::writable(account.address()));
 
-    accounts[1].write(InstructionAccount::readonly(mint.address()));
+    accounts[1].write(InstructionAccount::writable(mint.address()));
 
-    accounts[2].write(InstructionAccount::readonly(delegate.address()));
-
-    accounts[3].write(InstructionAccount::new(
+    accounts[2].write(InstructionAccount::new(
         authority.address(),
         false,
         multisig_signers.is_empty(),
     ));
 
-    for (account, signer) in accounts[4..expected_accounts]
+    for (account, signer) in accounts[3..expected_accounts]
         .iter_mut()
         .zip(multisig_signers.iter())
     {
