+++ programs/token/src/instructions/initialize_account2.rs
@@ -1,9 +1,11 @@
 use {
     crate::{
-        definitions::{
-            account_borrow_failed_error, invalid_argument_error, CpiWriter, TokenProgram,
+        instructions::{
+            account_borrow_failed_error,
+            invalid_argument_error, write_bytes, CpiWriter, UNINIT_BYTE, UNINIT_CPI_ACCOUNT,
+            UNINIT_INSTRUCTION_ACCOUNT,
         },
-        write_bytes, UNINIT_BYTE, UNINIT_CPI_ACCOUNT, UNINIT_INSTRUCTION_ACCOUNT,
+        TokenProgram,
     },
     core::{marker::PhantomData, mem::MaybeUninit, slice::from_raw_parts},
     solana_account_view::AccountView,
@@ -16,71 +18,76 @@ use {
 };
 
 /// The instruction discriminator.
-const DISCRIMINATOR: u8 = 0;
+const DISCRIMINATOR: u8 = 16;
 
 /// Expected number of accounts.
-const ACCOUNTS_LEN: usize = 2;
+const ACCOUNTS_LEN: usize = 3;
 
 /// Instruction data length:
 ///   - discriminator (1 byte)
-///   - decimals (1 byte)
-///   - mint authority (32 bytes)
-///   - freeze authority (33 bytes, optional)
-const MAX_DATA_LEN: usize = 67;
+///   - owner pubkey (32 bytes)
+const DATA_LEN: usize = 33;
 
-/// Initializes a new mint and optionally deposits all the newly minted
-/// tokens in an account.
-///
-/// The `InitializeMint` instruction requires no signers and MUST be
-/// included within the same Transaction as the system program's
-/// `CreateAccount` instruction that creates the account being initialized.
-/// Otherwise another party can acquire ownership of the uninitialized
-/// account.
+/// Like [`super::InitializeAccount`], but the owner pubkey is
+/// passed via instruction data rather than the accounts list. This
+/// variant may be preferable when using Cross Program Invocation from
+/// an instruction that does not need the owner's `AccountInfo`
+/// otherwise.
 ///
 /// Accounts expected by this instruction:
 ///
-///   0. `[writable]` The mint to initialize.
-///   1. `[]` Rent sysvar.
-pub struct InitializeMint<'account, 'address, Program: TokenProgram> {
-    /// The mint to initialize.
+///   0. `[writable]`  The account to initialize.
+///   1. `[]` The mint this account will be associated with.
+///   2. `[]` Rent sysvar.
+pub struct InitializeAccount2<'account, Program: TokenProgram> {
+    /// The account to initialize.
+    pub account: &'account AccountView,
+
+    /// The mint this account will be associated with.
     pub mint: &'account AccountView,
 
     /// Rent sysvar.
     pub rent_sysvar: &'account AccountView,
 
-    /// The number of base 10 digits to the right of the decimal place.
-    pub decimals: u8,
-
-    /// The authority/multisignature to mint tokens.
-    pub mint_authority: &'address Address,
-
-    /// The freeze authority/multisignature of the mint.
-    pub freeze_authority: Option<&'address Address>,
+    /// The new account's owner/multisignature.
+    pub owner: &'account Address,
 
     _program: PhantomData<Program>,
 }
 
-impl<'account, 'address, Program: TokenProgram> InitializeMint<'account, 'address, Program> {
+impl<'account, Program: TokenProgram> InitializeAccount2<'account, Program> {
+    /// The instruction discriminator.
+    pub const DISCRIMINATOR: u8 = DISCRIMINATOR;
+
+    /// Expected number of accounts.
+    pub const ACCOUNTS_LEN: usize = ACCOUNTS_LEN;
+
+    /// Instruction data length.
+    pub const DATA_LEN: usize = DATA_LEN;
+
     #[inline(always)]
     pub fn new(
+        account: &'account AccountView,
         mint: &'account AccountView,
         rent_sysvar: &'account AccountView,
-        decimals: u8,
-        mint_authority: &'address Address,
-        freeze_authority: Option<&'address Address>,
+        owner: &'account Address,
     ) -> Self {
         Self {
+            account,
             mint,
             rent_sysvar,
-            decimals,
-            mint_authority,
-            freeze_authority,
+            owner,
             _program: PhantomData,
         }
     }
 
     #[inline(always)]
     pub fn invoke(&self) -> ProgramResult {
+        self.invoke_with_program(&Program::ID)
+    }
+
+    #[inline(always)]
+    pub fn invoke_with_program(&self, program: &Address) -> ProgramResult {
         let mut instruction_accounts = [UNINIT_INSTRUCTION_ACCOUNT; ACCOUNTS_LEN];
         let written_instruction_accounts =
             self.write_instruction_accounts(&mut instruction_accounts)?;
@@ -88,20 +95,20 @@ impl<'account, 'address, Program: TokenProgram> InitializeMint<'account, 'addres
         let mut accounts = [UNINIT_CPI_ACCOUNT; ACCOUNTS_LEN];
         let written_accounts = self.write_accounts(&mut accounts)?;
 
-        let mut instruction_data = [UNINIT_BYTE; MAX_DATA_LEN];
+        let mut instruction_data = [UNINIT_BYTE; DATA_LEN];
         let written_instruction_data = self.write_instruction_data(&mut instruction_data)?;
 
         unsafe {
             invoke_unchecked(
                 &InstructionView {
-                    program_id: &Program::id(),
+                    program_id: program,
                     accounts: from_raw_parts(
                         instruction_accounts.as_ptr() as _,
                         written_instruction_accounts,
                     ),
                     data: from_raw_parts(instruction_data.as_ptr() as _, written_instruction_data),
                 },
-                from_raw_parts(accounts.as_ptr() as _, written_accounts),
+                from_raw_parts(accounts.as_ptr() as *const CpiAccount, written_accounts),
             );
         }
 
@@ -109,7 +116,7 @@ impl<'account, 'address, Program: TokenProgram> InitializeMint<'account, 'addres
     }
 }
 
-impl<Program: TokenProgram> CpiWriter for InitializeMint<'_, '_, Program> {
+impl<Program: TokenProgram> CpiWriter for InitializeAccount2<'_, Program> {
     #[inline(always)]
     fn write_accounts<'cpi>(
         &self,
@@ -118,7 +125,7 @@ impl<Program: TokenProgram> CpiWriter for InitializeMint<'_, '_, Program> {
     where
         Self: 'cpi,
     {
-        write_accounts(self.mint, self.rent_sysvar, accounts)
+        write_accounts(self.account, self.mint, self.rent_sysvar, accounts)
     }
 
     #[inline(always)]
@@ -129,46 +136,37 @@ impl<Program: TokenProgram> CpiWriter for InitializeMint<'_, '_, Program> {
     where
         Self: 'cpi,
     {
-        write_instruction_accounts(self.mint, self.rent_sysvar, accounts)
+        write_instruction_accounts(self.account, self.mint, self.rent_sysvar, accounts)
     }
 
     #[inline(always)]
     fn write_instruction_data(&self, data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
-        write_instruction_data(
-            self.decimals,
-            self.mint_authority,
-            self.freeze_authority,
-            data,
-        )
+        write_instruction_data(self.owner, data)
     }
 }
 
-impl<Program: TokenProgram> super::IntoBatch<Program> for InitializeMint<'_, '_, Program> {
+impl<Program: TokenProgram> super::batch::IntoBatch<Program> for InitializeAccount2<'_, Program> {
     #[inline(always)]
     fn into_batch<'account, 'state>(
         self,
-        batch: &mut super::Batch<'account, 'state, Program>,
+        batch: &mut super::batch::Batch<'account, 'state, Program>,
     ) -> ProgramResult
     where
         Self: 'account + 'state,
     {
         batch.push(
-            |accounts| write_accounts(self.mint, self.rent_sysvar, accounts),
-            |accounts| write_instruction_accounts(self.mint, self.rent_sysvar, accounts),
-            |data| {
-                write_instruction_data(
-                    self.decimals,
-                    self.mint_authority,
-                    self.freeze_authority,
-                    data,
-                )
+            |accounts| write_accounts(self.account, self.mint, self.rent_sysvar, accounts),
+            |accounts| {
+                write_instruction_accounts(self.account, self.mint, self.rent_sysvar, accounts)
             },
+            |data| write_instruction_data(self.owner, data),
         )
     }
 }
 
 #[inline(always)]
 fn write_accounts<'account, 'out>(
+    account: &'account AccountView,
     mint: &'account AccountView,
     rent_sysvar: &'account AccountView,
     accounts: &mut [MaybeUninit<CpiAccount<'out>>],
@@ -180,19 +178,22 @@ where
         return Err(invalid_argument_error());
     }
 
-    if mint.is_borrowed() {
+    if account.is_borrowed() {
         return Err(account_borrow_failed_error());
     }
 
-    CpiAccount::init_from_account_view(mint, &mut accounts[0]);
+    CpiAccount::init_from_account_view(account, &mut accounts[0]);
 
-    CpiAccount::init_from_account_view(rent_sysvar, &mut accounts[1]);
+    CpiAccount::init_from_account_view(mint, &mut accounts[1]);
+
+    CpiAccount::init_from_account_view(rent_sysvar, &mut accounts[2]);
 
     Ok(ACCOUNTS_LEN)
 }
 
 #[inline(always)]
 fn write_instruction_accounts<'account, 'out>(
+    account: &'account AccountView,
     mint: &'account AccountView,
     rent_sysvar: &'account AccountView,
     accounts: &mut [MaybeUninit<InstructionAccount<'out>>],
@@ -204,39 +205,27 @@ where
         return Err(invalid_argument_error());
     }
 
-    accounts[0].write(InstructionAccount::writable(mint.address()));
+    accounts[0].write(InstructionAccount::writable(account.address()));
 
-    accounts[1].write(InstructionAccount::readonly(rent_sysvar.address()));
+    accounts[1].write(InstructionAccount::readonly(mint.address()));
+
+    accounts[2].write(InstructionAccount::readonly(rent_sysvar.address()));
 
     Ok(ACCOUNTS_LEN)
 }
 
 #[inline(always)]
 fn write_instruction_data(
-    decimals: u8,
-    mint_authority: &Address,
-    freeze_authority: Option<&Address>,
+    owner: &Address,
     data: &mut [MaybeUninit<u8>],
 ) -> Result<usize, ProgramError> {
-    if data.len() < MAX_DATA_LEN {
+    if data.len() < DATA_LEN {
         return Err(invalid_argument_error());
     }
 
     data[0].write(DISCRIMINATOR);
 
-    data[1].write(decimals);
+    write_bytes(&mut data[1..DATA_LEN], owner.as_array());
 
-    write_bytes(&mut data[2..34], mint_authority.as_array());
-
-    if let Some(freeze_authority) = freeze_authority {
-        data[34].write(1);
-
-        write_bytes(&mut data[35..MAX_DATA_LEN], freeze_authority.as_array());
-
-        Ok(MAX_DATA_LEN)
-    } else {
-        data[34].write(0);
-
-        Ok(35)
-    }
+    Ok(DATA_LEN)
 }
