+++ programs/token/src/instructions/initialize_mint.rs
@@ -1,9 +1,11 @@
 use {
     crate::{
-        definitions::{
-            account_borrow_failed_error, invalid_argument_error, CpiWriter, TokenProgram,
+        instructions::{
+            account_borrow_failed_error,
+            invalid_argument_error, write_bytes, CpiWriter, UNINIT_BYTE, UNINIT_CPI_ACCOUNT,
+            UNINIT_INSTRUCTION_ACCOUNT,
         },
-        write_bytes, UNINIT_BYTE, UNINIT_CPI_ACCOUNT, UNINIT_INSTRUCTION_ACCOUNT,
+        TokenProgram,
     },
     core::{marker::PhantomData, mem::MaybeUninit, slice::from_raw_parts},
     solana_account_view::AccountView,
@@ -16,10 +18,10 @@ use {
 };
 
 /// The instruction discriminator.
-const DISCRIMINATOR: u8 = 20;
+const DISCRIMINATOR: u8 = 0;
 
 /// Expected number of accounts.
-const ACCOUNTS_LEN: usize = 1;
+const ACCOUNTS_LEN: usize = 2;
 
 /// Instruction data length:
 ///   - discriminator (1 byte)
@@ -28,15 +30,25 @@ const ACCOUNTS_LEN: usize = 1;
 ///   - freeze authority (33 bytes, optional)
 const MAX_DATA_LEN: usize = 67;
 
-/// Like [`super::InitializeMint`], but does not require the Rent
-/// sysvar to be provided
+/// Initializes a new mint and optionally deposits all the newly minted
+/// tokens in an account.
+///
+/// The `InitializeMint` instruction requires no signers and MUST be
+/// included within the same Transaction as the system program's
+/// `CreateAccount` instruction that creates the account being initialized.
+/// Otherwise another party can acquire ownership of the uninitialized
+/// account.
 ///
 /// Accounts expected by this instruction:
 ///
 ///   0. `[writable]` The mint to initialize.
-pub struct InitializeMint2<'account, 'address, Program: TokenProgram> {
+///   1. `[]` Rent sysvar.
+pub struct InitializeMint<'account, 'address, Program: TokenProgram> {
     /// The mint to initialize.
     pub mint: &'account AccountView,
+
+    /// Rent sysvar.
+    pub rent_sysvar: &'account AccountView,
 
     /// The number of base 10 digits to the right of the decimal place.
     pub decimals: u8,
@@ -50,16 +62,27 @@ pub struct InitializeMint2<'account, 'address, Program: TokenProgram> {
     _program: PhantomData<Program>,
 }
 
-impl<'account, 'address, Program: TokenProgram> InitializeMint2<'account, 'address, Program> {
+impl<'account, 'address, Program: TokenProgram> InitializeMint<'account, 'address, Program> {
+    /// The instruction discriminator.
+    pub const DISCRIMINATOR: u8 = DISCRIMINATOR;
+
+    /// Expected number of accounts.
+    pub const ACCOUNTS_LEN: usize = ACCOUNTS_LEN;
+
+    /// Maximum instruction data length.
+    pub const MAX_DATA_LEN: usize = MAX_DATA_LEN;
+
     #[inline(always)]
     pub fn new(
         mint: &'account AccountView,
+        rent_sysvar: &'account AccountView,
         decimals: u8,
         mint_authority: &'address Address,
         freeze_authority: Option<&'address Address>,
     ) -> Self {
         Self {
             mint,
+            rent_sysvar,
             decimals,
             mint_authority,
             freeze_authority,
@@ -69,6 +92,11 @@ impl<'account, 'address, Program: TokenProgram> InitializeMint2<'account, 'addre
 
     #[inline(always)]
     pub fn invoke(&self) -> ProgramResult {
+        self.invoke_with_program(&Program::ID)
+    }
+
+    #[inline(always)]
+    pub fn invoke_with_program(&self, program: &Address) -> ProgramResult {
         let mut instruction_accounts = [UNINIT_INSTRUCTION_ACCOUNT; ACCOUNTS_LEN];
         let written_instruction_accounts =
             self.write_instruction_accounts(&mut instruction_accounts)?;
@@ -82,7 +110,7 @@ impl<'account, 'address, Program: TokenProgram> InitializeMint2<'account, 'addre
         unsafe {
             invoke_unchecked(
                 &InstructionView {
-                    program_id: &Program::id(),
+                    program_id: program,
                     accounts: from_raw_parts(
                         instruction_accounts.as_ptr() as _,
                         written_instruction_accounts,
@@ -97,7 +125,7 @@ impl<'account, 'address, Program: TokenProgram> InitializeMint2<'account, 'addre
     }
 }
 
-impl<Program: TokenProgram> CpiWriter for InitializeMint2<'_, '_, Program> {
+impl<Program: TokenProgram> CpiWriter for InitializeMint<'_, '_, Program> {
     #[inline(always)]
     fn write_accounts<'cpi>(
         &self,
@@ -106,7 +134,7 @@ impl<Program: TokenProgram> CpiWriter for InitializeMint2<'_, '_, Program> {
     where
         Self: 'cpi,
     {
-        write_accounts(self.mint, accounts)
+        write_accounts(self.mint, self.rent_sysvar, accounts)
     }
 
     #[inline(always)]
@@ -117,7 +145,7 @@ impl<Program: TokenProgram> CpiWriter for InitializeMint2<'_, '_, Program> {
     where
         Self: 'cpi,
     {
-        write_instruction_accounts(self.mint, accounts)
+        write_instruction_accounts(self.mint, self.rent_sysvar, accounts)
     }
 
     #[inline(always)]
@@ -131,18 +159,18 @@ impl<Program: TokenProgram> CpiWriter for InitializeMint2<'_, '_, Program> {
     }
 }
 
-impl<Program: TokenProgram> super::IntoBatch<Program> for InitializeMint2<'_, '_, Program> {
+impl<Program: TokenProgram> super::batch::IntoBatch<Program> for InitializeMint<'_, '_, Program> {
     #[inline(always)]
     fn into_batch<'account, 'state>(
         self,
-        batch: &mut super::Batch<'account, 'state, Program>,
+        batch: &mut super::batch::Batch<'account, 'state, Program>,
     ) -> ProgramResult
     where
         Self: 'account + 'state,
     {
         batch.push(
-            |accounts| write_accounts(self.mint, accounts),
-            |accounts| write_instruction_accounts(self.mint, accounts),
+            |accounts| write_accounts(self.mint, self.rent_sysvar, accounts),
+            |accounts| write_instruction_accounts(self.mint, self.rent_sysvar, accounts),
             |data| {
                 write_instruction_data(
                     self.decimals,
@@ -158,6 +186,7 @@ impl<Program: TokenProgram> super::IntoBatch<Program> for InitializeMint2<'_, '_
 #[inline(always)]
 fn write_accounts<'account, 'out>(
     mint: &'account AccountView,
+    rent_sysvar: &'account AccountView,
     accounts: &mut [MaybeUninit<CpiAccount<'out>>],
 ) -> Result<usize, ProgramError>
 where
@@ -173,12 +202,15 @@ where
 
     CpiAccount::init_from_account_view(mint, &mut accounts[0]);
 
+    CpiAccount::init_from_account_view(rent_sysvar, &mut accounts[1]);
+
     Ok(ACCOUNTS_LEN)
 }
 
 #[inline(always)]
 fn write_instruction_accounts<'account, 'out>(
     mint: &'account AccountView,
+    rent_sysvar: &'account AccountView,
     accounts: &mut [MaybeUninit<InstructionAccount<'out>>],
 ) -> Result<usize, ProgramError>
 where
@@ -189,6 +221,8 @@ where
     }
 
     accounts[0].write(InstructionAccount::writable(mint.address()));
+
+    accounts[1].write(InstructionAccount::readonly(rent_sysvar.address()));
 
     Ok(ACCOUNTS_LEN)
 }
@@ -210,10 +244,10 @@ fn write_instruction_data(
 
     write_bytes(&mut data[2..34], mint_authority.as_array());
 
-    if let Some(freeze_auth) = freeze_authority {
+    if let Some(freeze_authority) = freeze_authority {
         data[34].write(1);
 
-        write_bytes(&mut data[35..MAX_DATA_LEN], freeze_auth.as_array());
+        write_bytes(&mut data[35..MAX_DATA_LEN], freeze_authority.as_array());
 
         Ok(MAX_DATA_LEN)
     } else {
