+++ programs/token/src/instructions/initialize_mint2.rs
@@ -1,9 +1,11 @@
 use {
     crate::{
-        definitions::{
-            account_borrow_failed_error, invalid_argument_error, CpiWriter, TokenProgram,
+        instructions::{
+            account_borrow_failed_error,
+            invalid_argument_error, write_bytes, CpiWriter, UNINIT_BYTE, UNINIT_CPI_ACCOUNT,
+            UNINIT_INSTRUCTION_ACCOUNT,
         },
-        write_bytes, UNINIT_BYTE, UNINIT_CPI_ACCOUNT, UNINIT_INSTRUCTION_ACCOUNT,
+        TokenProgram,
     },
     core::{marker::PhantomData, mem::MaybeUninit, slice::from_raw_parts},
     solana_account_view::AccountView,
@@ -16,62 +18,73 @@ use {
 };
 
 /// The instruction discriminator.
-const DISCRIMINATOR: u8 = 16;
+const DISCRIMINATOR: u8 = 20;
 
 /// Expected number of accounts.
-const ACCOUNTS_LEN: usize = 3;
+const ACCOUNTS_LEN: usize = 1;
 
 /// Instruction data length:
 ///   - discriminator (1 byte)
-///   - owner pubkey (32 bytes)
-const DATA_LEN: usize = 33;
+///   - decimals (1 byte)
+///   - mint authority (32 bytes)
+///   - freeze authority (33 bytes, optional)
+const MAX_DATA_LEN: usize = 67;
 
-/// Like [`super::InitializeAccount`], but the owner pubkey is
-/// passed via instruction data rather than the accounts list. This
-/// variant may be preferable when using Cross Program Invocation from
-/// an instruction that does not need the owner's `AccountInfo`
-/// otherwise.
+/// Like [`super::InitializeMint`], but does not require the Rent
+/// sysvar to be provided
 ///
 /// Accounts expected by this instruction:
 ///
-///   0. `[writable]`  The account to initialize.
-///   1. `[]` The mint this account will be associated with.
-///   2. `[]` Rent sysvar.
-pub struct InitializeAccount2<'account, Program: TokenProgram> {
-    /// The account to initialize.
-    pub account: &'account AccountView,
-
-    /// The mint this account will be associated with.
+///   0. `[writable]` The mint to initialize.
+pub struct InitializeMint2<'account, 'address, Program: TokenProgram> {
+    /// The mint to initialize.
     pub mint: &'account AccountView,
 
-    /// Rent sysvar.
-    pub rent_sysvar: &'account AccountView,
+    /// The number of base 10 digits to the right of the decimal place.
+    pub decimals: u8,
 
-    /// The new account's owner/multisignature.
-    pub owner: &'account Address,
+    /// The authority/multisignature to mint tokens.
+    pub mint_authority: &'address Address,
+
+    /// The freeze authority/multisignature of the mint.
+    pub freeze_authority: Option<&'address Address>,
 
     _program: PhantomData<Program>,
 }
 
-impl<'account, Program: TokenProgram> InitializeAccount2<'account, Program> {
+impl<'account, 'address, Program: TokenProgram> InitializeMint2<'account, 'address, Program> {
+    /// The instruction discriminator.
+    pub const DISCRIMINATOR: u8 = DISCRIMINATOR;
+
+    /// Expected number of accounts.
+    pub const ACCOUNTS_LEN: usize = ACCOUNTS_LEN;
+
+    /// Maximum instruction data length.
+    pub const MAX_DATA_LEN: usize = MAX_DATA_LEN;
+
     #[inline(always)]
     pub fn new(
-        account: &'account AccountView,
         mint: &'account AccountView,
-        rent_sysvar: &'account AccountView,
-        owner: &'account Address,
+        decimals: u8,
+        mint_authority: &'address Address,
+        freeze_authority: Option<&'address Address>,
     ) -> Self {
         Self {
-            account,
             mint,
-            rent_sysvar,
-            owner,
+            decimals,
+            mint_authority,
+            freeze_authority,
             _program: PhantomData,
         }
     }
 
     #[inline(always)]
     pub fn invoke(&self) -> ProgramResult {
+        self.invoke_with_program(&Program::ID)
+    }
+
+    #[inline(always)]
+    pub fn invoke_with_program(&self, program: &Address) -> ProgramResult {
         let mut instruction_accounts = [UNINIT_INSTRUCTION_ACCOUNT; ACCOUNTS_LEN];
         let written_instruction_accounts =
             self.write_instruction_accounts(&mut instruction_accounts)?;
@@ -79,20 +92,20 @@ impl<'account, Program: TokenProgram> InitializeAccount2<'account, Program> {
         let mut accounts = [UNINIT_CPI_ACCOUNT; ACCOUNTS_LEN];
         let written_accounts = self.write_accounts(&mut accounts)?;
 
-        let mut instruction_data = [UNINIT_BYTE; DATA_LEN];
+        let mut instruction_data = [UNINIT_BYTE; MAX_DATA_LEN];
         let written_instruction_data = self.write_instruction_data(&mut instruction_data)?;
 
         unsafe {
             invoke_unchecked(
                 &InstructionView {
-                    program_id: &Program::id(),
+                    program_id: program,
                     accounts: from_raw_parts(
                         instruction_accounts.as_ptr() as _,
                         written_instruction_accounts,
                     ),
                     data: from_raw_parts(instruction_data.as_ptr() as _, written_instruction_data),
                 },
-                from_raw_parts(accounts.as_ptr() as *const CpiAccount, written_accounts),
+                from_raw_parts(accounts.as_ptr() as _, written_accounts),
             );
         }
 
@@ -100,7 +113,7 @@ impl<'account, Program: TokenProgram> InitializeAccount2<'account, Program> {
     }
 }
 
-impl<Program: TokenProgram> CpiWriter for InitializeAccount2<'_, Program> {
+impl<Program: TokenProgram> CpiWriter for InitializeMint2<'_, '_, Program> {
     #[inline(always)]
     fn write_accounts<'cpi>(
         &self,
@@ -109,7 +122,7 @@ impl<Program: TokenProgram> CpiWriter for InitializeAccount2<'_, Program> {
     where
         Self: 'cpi,
     {
-        write_accounts(self.account, self.mint, self.rent_sysvar, accounts)
+        write_accounts(self.mint, accounts)
     }
 
     #[inline(always)]
@@ -120,39 +133,47 @@ impl<Program: TokenProgram> CpiWriter for InitializeAccount2<'_, Program> {
     where
         Self: 'cpi,
     {
-        write_instruction_accounts(self.account, self.mint, self.rent_sysvar, accounts)
+        write_instruction_accounts(self.mint, accounts)
     }
 
     #[inline(always)]
     fn write_instruction_data(&self, data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
-        write_instruction_data(self.owner, data)
+        write_instruction_data(
+            self.decimals,
+            self.mint_authority,
+            self.freeze_authority,
+            data,
+        )
     }
 }
 
-impl<Program: TokenProgram> super::IntoBatch<Program> for InitializeAccount2<'_, Program> {
+impl<Program: TokenProgram> super::batch::IntoBatch<Program> for InitializeMint2<'_, '_, Program> {
     #[inline(always)]
     fn into_batch<'account, 'state>(
         self,
-        batch: &mut super::Batch<'account, 'state, Program>,
+        batch: &mut super::batch::Batch<'account, 'state, Program>,
     ) -> ProgramResult
     where
         Self: 'account + 'state,
     {
         batch.push(
-            |accounts| write_accounts(self.account, self.mint, self.rent_sysvar, accounts),
-            |accounts| {
-                write_instruction_accounts(self.account, self.mint, self.rent_sysvar, accounts)
+            |accounts| write_accounts(self.mint, accounts),
+            |accounts| write_instruction_accounts(self.mint, accounts),
+            |data| {
+                write_instruction_data(
+                    self.decimals,
+                    self.mint_authority,
+                    self.freeze_authority,
+                    data,
+                )
             },
-            |data| write_instruction_data(self.owner, data),
         )
     }
 }
 
 #[inline(always)]
 fn write_accounts<'account, 'out>(
-    account: &'account AccountView,
     mint: &'account AccountView,
-    rent_sysvar: &'account AccountView,
     accounts: &mut [MaybeUninit<CpiAccount<'out>>],
 ) -> Result<usize, ProgramError>
 where
@@ -162,24 +183,18 @@ where
         return Err(invalid_argument_error());
     }
 
-    if account.is_borrowed() {
+    if mint.is_borrowed() {
         return Err(account_borrow_failed_error());
     }
 
-    CpiAccount::init_from_account_view(account, &mut accounts[0]);
-
-    CpiAccount::init_from_account_view(mint, &mut accounts[1]);
-
-    CpiAccount::init_from_account_view(rent_sysvar, &mut accounts[2]);
+    CpiAccount::init_from_account_view(mint, &mut accounts[0]);
 
     Ok(ACCOUNTS_LEN)
 }
 
 #[inline(always)]
 fn write_instruction_accounts<'account, 'out>(
-    account: &'account AccountView,
     mint: &'account AccountView,
-    rent_sysvar: &'account AccountView,
     accounts: &mut [MaybeUninit<InstructionAccount<'out>>],
 ) -> Result<usize, ProgramError>
 where
@@ -189,27 +204,37 @@ where
         return Err(invalid_argument_error());
     }
 
-    accounts[0].write(InstructionAccount::writable(account.address()));
-
-    accounts[1].write(InstructionAccount::readonly(mint.address()));
-
-    accounts[2].write(InstructionAccount::readonly(rent_sysvar.address()));
+    accounts[0].write(InstructionAccount::writable(mint.address()));
 
     Ok(ACCOUNTS_LEN)
 }
 
 #[inline(always)]
 fn write_instruction_data(
-    owner: &Address,
+    decimals: u8,
+    mint_authority: &Address,
+    freeze_authority: Option<&Address>,
     data: &mut [MaybeUninit<u8>],
 ) -> Result<usize, ProgramError> {
-    if data.len() < DATA_LEN {
+    if data.len() < MAX_DATA_LEN {
         return Err(invalid_argument_error());
     }
 
     data[0].write(DISCRIMINATOR);
 
-    write_bytes(&mut data[1..DATA_LEN], owner.as_array());
+    data[1].write(decimals);
 
-    Ok(DATA_LEN)
+    write_bytes(&mut data[2..34], mint_authority.as_array());
+
+    if let Some(freeze_auth) = freeze_authority {
+        data[34].write(1);
+
+        write_bytes(&mut data[35..MAX_DATA_LEN], freeze_auth.as_array());
+
+        Ok(MAX_DATA_LEN)
+    } else {
+        data[34].write(0);
+
+        Ok(35)
+    }
 }
