+++ programs/token/src/instructions/mint_to.rs
@@ -1,13 +1,15 @@
 use {
     crate::{
-        definitions::{
-            account_borrow_failed_error, invalid_argument_error, CpiWriter, TokenProgram,
-            MAX_MULTISIG_SIGNERS,
+        instructions::{
+            account_borrow_failed_error, initialize_multisig::MAX_MULTISIG_SIGNERS,
+            invalid_argument_error, write_bytes, CpiWriter, UNINIT_BYTE, UNINIT_CPI_ACCOUNT,
+            UNINIT_INSTRUCTION_ACCOUNT,
         },
-        write_bytes, UNINIT_BYTE, UNINIT_CPI_ACCOUNT, UNINIT_INSTRUCTION_ACCOUNT,
+        TokenProgram,
     },
     core::{marker::PhantomData, mem::MaybeUninit, slice::from_raw_parts},
     solana_account_view::AccountView,
+    solana_address::Address,
     solana_instruction_view::{
         cpi::{invoke_signed_unchecked, CpiAccount, Signer},
         InstructionAccount, InstructionView,
@@ -16,7 +18,7 @@ use {
 };
 
 /// The instruction discriminator.
-const DISCRIMINATOR: u8 = 4;
+const DISCRIMINATOR: u8 = 7;
 
 /// Maximum number of accounts expected by this instruction.
 ///
@@ -27,73 +29,82 @@ const MAX_ACCOUNTS_LEN: usize = 3 + MAX_MULTISIG_SIGNERS;
 
 /// Instruction data length:
 ///   - discriminator (1 byte)
-///   - amount (8 bytes)
+///   - amount to mint (8 bytes)
 const DATA_LEN: usize = 9;
 
-/// Approves a delegate.  A delegate is given the authority over tokens on
-/// behalf of the source account's owner.
+/// Mints new tokens to an account.  The native mint does not support
+/// minting.
 ///
 /// Accounts expected by this instruction:
 ///
-///   * Single owner
-///   0. `[writable]` The source account.
-///   1. `[]` The delegate.
-///   2. `[signer]` The source account owner.
+///   * Single authority
+///   0. `[writable]` The mint.
+///   1. `[writable]` The account to mint tokens to.
+///   2. `[signer]` The mint's minting authority.
 ///
-///   * Multisignature owner
-///   0. `[writable]` The source account.
-///   1. `[]` The delegate.
-///   2. `[]` The source account's multisignature owner.
+///   * Multisignature authority
+///   0. `[writable]` The mint.
+///   1. `[writable]` The account to mint tokens to.
+///   2. `[]` The mint's multisignature mint-tokens authority.
 ///   3. `..+M` `[signer]` M signer accounts.
-pub struct Approve<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram> {
-    /// The source account.
-    pub source: &'account AccountView,
+pub struct MintTo<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram> {
+    /// The mint.
+    pub mint: &'account AccountView,
 
-    /// The delegate.
-    pub delegate: &'account AccountView,
+    /// The account to mint tokens to.
+    pub account: &'account AccountView,
 
-    /// The source account owner.
-    pub authority: &'account AccountView,
+    /// The account to mint tokens to.
+    pub mint_authority: &'account AccountView,
 
     /// Multisignature signers.
     pub multisig_signers: &'multisig [MultisigSigner],
 
-    /// The amount of tokens the delegate is approved for.
+    /// The amount of new tokens to mint.
     pub amount: u64,
 
     _program: PhantomData<Program>,
 }
 
-impl<'account, Program: TokenProgram> Approve<'account, '_, &'account AccountView, Program> {
-    /// Creates a new `Approve` instruction with a single owner authority.
+impl<'account, Program: TokenProgram> MintTo<'account, '_, &'account AccountView, Program> {
+    /// The instruction discriminator.
+    pub const DISCRIMINATOR: u8 = DISCRIMINATOR;
+
+    /// Maximum number of accounts expected by this instruction.
+    pub const MAX_ACCOUNTS_LEN: usize = MAX_ACCOUNTS_LEN;
+
+    /// Instruction data length.
+    pub const DATA_LEN: usize = DATA_LEN;
+
+    /// Creates a new `MintTo` instruction with a single mint authority.
     #[inline(always)]
     pub fn new(
-        source: &'account AccountView,
-        delegate: &'account AccountView,
-        authority: &'account AccountView,
+        mint: &'account AccountView,
+        account: &'account AccountView,
+        mint_authority: &'account AccountView,
         amount: u64,
     ) -> Self {
-        Self::with_multisig_signers(source, delegate, authority, amount, &[])
+        Self::with_multisig_signers(mint, account, mint_authority, amount, &[])
     }
 }
 
 impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram>
-    Approve<'account, 'multisig, MultisigSigner, Program>
+    MintTo<'account, 'multisig, MultisigSigner, Program>
 {
-    /// Creates a new `Approve` instruction with a
-    /// multisignature owner authority and signer accounts.
+    /// Creates a new `MintTo` instruction with a
+    /// multisignature mint authority and signer accounts.
     #[inline(always)]
     pub fn with_multisig_signers(
-        source: &'account AccountView,
-        delegate: &'account AccountView,
-        authority: &'account AccountView,
+        mint: &'account AccountView,
+        account: &'account AccountView,
+        mint_authority: &'account AccountView,
         amount: u64,
         multisig_signers: &'multisig [MultisigSigner],
     ) -> Self {
         Self {
-            source,
-            delegate,
-            authority,
+            mint,
+            account,
+            mint_authority,
             multisig_signers,
             amount,
             _program: PhantomData,
@@ -102,11 +113,25 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
 
     #[inline(always)]
     pub fn invoke(&self) -> ProgramResult {
-        self.invoke_signed(&[])
+        self.invoke_with_program(&Program::ID)
     }
 
     #[inline(always)]
     pub fn invoke_signed(&self, signers: &[Signer]) -> ProgramResult {
+        self.invoke_signed_with_program(signers, &Program::ID)
+    }
+
+    #[inline(always)]
+    pub fn invoke_with_program(&self, program: &Address) -> ProgramResult {
+        self.invoke_signed_with_program(&[], program)
+    }
+
+    #[inline(always)]
+    pub fn invoke_signed_with_program(
+        &self,
+        signers: &[Signer],
+        program: &Address,
+    ) -> ProgramResult {
         if self.multisig_signers.len() > MAX_MULTISIG_SIGNERS {
             Err(ProgramError::InvalidArgument)?;
         }
@@ -124,7 +149,7 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
         unsafe {
             invoke_signed_unchecked(
                 &InstructionView {
-                    program_id: &Program::id(),
+                    program_id: program,
                     accounts: from_raw_parts(
                         instruction_accounts.as_ptr() as _,
                         written_instruction_accounts,
@@ -141,7 +166,7 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
 }
 
 impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
-    for Approve<'_, '_, MultisigSigner, Program>
+    for MintTo<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn write_accounts<'cpi>(
@@ -152,9 +177,9 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
         Self: 'cpi,
     {
         write_accounts(
-            self.source,
-            self.delegate,
-            self.authority,
+            self.mint,
+            self.account,
+            self.mint_authority,
             self.multisig_signers,
             accounts,
         )
@@ -169,9 +194,9 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
         Self: 'cpi,
     {
         write_instruction_accounts(
-            self.source,
-            self.delegate,
-            self.authority,
+            self.mint,
+            self.account,
+            self.mint_authority,
             self.multisig_signers,
             accounts,
         )
@@ -183,41 +208,46 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
     }
 }
 
-impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch<Program>
-    for Approve<'_, '_, MultisigSigner, Program>
+impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::batch::IntoBatch<Program>
+    for MintTo<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn into_batch<'account, 'state>(
         self,
-        batch: &mut super::Batch<'account, 'state, Program>,
+        batch: &mut super::batch::Batch<'account, 'state, Program>,
     ) -> ProgramResult
     where
         Self: 'account + 'state,
     {
-        let Self {
-            source,
-            delegate,
-            authority,
-            multisig_signers,
-            amount,
-            ..
-        } = self;
-
         batch.push(
-            |accounts| write_accounts(source, delegate, authority, multisig_signers, accounts),
             |accounts| {
-                write_instruction_accounts(source, delegate, authority, multisig_signers, accounts)
+                write_accounts(
+                    self.mint,
+                    self.account,
+                    self.mint_authority,
+                    self.multisig_signers,
+                    accounts,
+                )
             },
-            |data| write_instruction_data(amount, data),
+            |accounts| {
+                write_instruction_accounts(
+                    self.mint,
+                    self.account,
+                    self.mint_authority,
+                    self.multisig_signers,
+                    accounts,
+                )
+            },
+            |data| write_instruction_data(self.amount, data),
         )
     }
 }
 
 #[inline(always)]
 fn write_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
-    source: &'account AccountView,
-    delegate: &'account AccountView,
-    authority: &'account AccountView,
+    mint: &'account AccountView,
+    account: &'account AccountView,
+    mint_authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<CpiAccount<'out>>],
 ) -> Result<usize, ProgramError>
@@ -231,15 +261,15 @@ where
         return Err(invalid_argument_error());
     }
 
-    if source.is_borrowed() {
+    if mint.is_borrowed() | account.is_borrowed() {
         return Err(account_borrow_failed_error());
     }
 
-    CpiAccount::init_from_account_view(source, &mut accounts[0]);
+    CpiAccount::init_from_account_view(mint, &mut accounts[0]);
 
-    CpiAccount::init_from_account_view(delegate, &mut accounts[1]);
+    CpiAccount::init_from_account_view(account, &mut accounts[1]);
 
-    CpiAccount::init_from_account_view(authority, &mut accounts[2]);
+    CpiAccount::init_from_account_view(mint_authority, &mut accounts[2]);
 
     for (account, signer) in accounts[3..expected_accounts]
         .iter_mut()
@@ -253,9 +283,9 @@ where
 
 #[inline(always)]
 fn write_instruction_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
-    source: &'account AccountView,
-    delegate: &'account AccountView,
-    authority: &'account AccountView,
+    mint: &'account AccountView,
+    account: &'account AccountView,
+    mint_authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<InstructionAccount<'out>>],
 ) -> Result<usize, ProgramError>
@@ -269,12 +299,12 @@ where
         return Err(invalid_argument_error());
     }
 
-    accounts[0].write(InstructionAccount::writable(source.address()));
+    accounts[0].write(InstructionAccount::writable(mint.address()));
 
-    accounts[1].write(InstructionAccount::readonly(delegate.address()));
+    accounts[1].write(InstructionAccount::writable(account.address()));
 
     accounts[2].write(InstructionAccount::new(
-        authority.address(),
+        mint_authority.address(),
         false,
         multisig_signers.is_empty(),
     ));
