+++ programs/token/src/instructions/mod.rs
@@ -1,18 +1,50 @@
-use crate::definitions::TokenProgram;
-use solana_address::Address;
-
 #[cfg(feature = "alloc")]
-pub use crate::definitions::{BatchState, IntoBatch};
+pub use crate::instructions::batch::{BatchState, IntoBatch};
+use {
+    crate::Program,
+    core::mem::MaybeUninit,
+    solana_instruction_view::{cpi::CpiAccount, InstructionAccount},
+    solana_program_error::ProgramError,
+};
 
-/// Represent the SPL Token program.
-pub struct Program;
+pub mod amount_to_ui_amount;
+pub mod approve;
+pub mod approve_checked;
+pub mod batch;
+pub mod burn;
+pub mod burn_checked;
+pub mod close_account;
+pub mod freeze_account;
+pub mod get_account_data_size;
+pub mod initialize_account;
+pub mod initialize_account2;
+pub mod initialize_account3;
+pub mod initialize_immutable_owner;
+pub mod initialize_mint;
+pub mod initialize_mint2;
+pub mod initialize_multisig;
+pub mod initialize_multisig2;
+pub mod mint_to;
+pub mod mint_to_checked;
+pub mod revoke;
+pub mod set_authority;
+pub mod sync_native;
+pub mod thaw_account;
+pub mod transfer;
+pub mod transfer_checked;
+pub mod ui_amount_to_amount;
+pub mod unwrap_lamports;
+pub mod withdraw_excess_lamports;
 
-impl TokenProgram for Program {
-    #[inline(always)]
-    fn id() -> Address {
-        crate::ID
-    }
-}
+/// Constant for an uninitialized byte.
+const UNINIT_BYTE: MaybeUninit<u8> = MaybeUninit::<u8>::uninit();
+
+/// Constant for an uninitialized `CpiAccount`.
+const UNINIT_CPI_ACCOUNT: MaybeUninit<CpiAccount> = MaybeUninit::<CpiAccount>::uninit();
+
+/// Constant for an uninitialized `InstructionAccount`.
+const UNINIT_INSTRUCTION_ACCOUNT: MaybeUninit<InstructionAccount> =
+    MaybeUninit::<InstructionAccount>::uninit();
 
 /// Convert an Amount of tokens to a `UiAmount` string, using the given
 /// mint.
@@ -25,7 +57,7 @@ impl TokenProgram for Program {
 /// Accounts expected by this instruction:
 ///
 ///   0. `[]` The mint to calculate for.
-pub type AmountToUiAmount<'account> = crate::definitions::AmountToUiAmount<'account, Program>;
+pub type AmountToUiAmount<'account> = amount_to_ui_amount::AmountToUiAmount<'account, Program>;
 
 /// Approves a delegate.  A delegate is given the authority over tokens on
 /// behalf of the source account's owner.
@@ -43,7 +75,7 @@ pub type AmountToUiAmount<'account> = crate::definitions::AmountToUiAmount<'acco
 ///   2. `[]` The source account's multisignature owner.
 ///   3. `..+M` `[signer]` M signer accounts.
 pub type Approve<'account, 'multisig, MultisigSigner> =
-    crate::definitions::Approve<'account, 'multisig, MultisigSigner, Program>;
+    approve::Approve<'account, 'multisig, MultisigSigner, Program>;
 
 /// Approves a delegate.  A delegate is given the authority over tokens on
 /// behalf of the source account's owner.
@@ -67,11 +99,11 @@ pub type Approve<'account, 'multisig, MultisigSigner> =
 ///   3. `[]` The source account's multisignature owner.
 ///   4. `..+M` `[signer]` M signer accounts.
 pub type ApproveChecked<'account, 'multisig, MultisigSigner> =
-    crate::definitions::ApproveChecked<'account, 'multisig, MultisigSigner, Program>;
+    approve_checked::ApproveChecked<'account, 'multisig, MultisigSigner, Program>;
 
 /// A collection of instructions that can be serialized into a token `Batch`
 /// instruction.
-pub type Batch<'account, 'state> = crate::definitions::Batch<'account, 'state, Program>;
+pub type Batch<'account, 'state> = batch::Batch<'account, 'state, Program>;
 
 /// Burns tokens by removing them from an account.  `Burn` does not support
 /// accounts associated with the native mint, use `CloseAccount` instead.
@@ -89,7 +121,7 @@ pub type Batch<'account, 'state> = crate::definitions::Batch<'account, 'state, P
 ///   2. `[]` The account's multisignature owner/delegate.
 ///   3. `..+M` `[signer]` M signer accounts.
 pub type Burn<'account, 'multisig, MultisigSigner> =
-    crate::definitions::Burn<'account, 'multisig, MultisigSigner, Program>;
+    burn::Burn<'account, 'multisig, MultisigSigner, Program>;
 
 /// Burns tokens by removing them from an account.
 /// [`BurnChecked`] does not support accounts
@@ -112,7 +144,7 @@ pub type Burn<'account, 'multisig, MultisigSigner> =
 ///   2. `[]` The account's multisignature owner/delegate.
 ///   3. `..+M` `[signer]` M signer accounts.
 pub type BurnChecked<'account, 'multisig, MultisigSigner> =
-    crate::definitions::BurnChecked<'account, 'multisig, MultisigSigner, Program>;
+    burn_checked::BurnChecked<'account, 'multisig, MultisigSigner, Program>;
 
 /// Close an account by transferring all its SOL to the destination account.
 /// Non-native accounts may only be closed if its token amount is zero.
@@ -130,7 +162,7 @@ pub type BurnChecked<'account, 'multisig, MultisigSigner> =
 ///   2. `[]` The account's multisignature owner.
 ///   3. `..+M` `[signer]` M signer accounts.
 pub type CloseAccount<'account, 'multisig, MultisigSigner> =
-    crate::definitions::CloseAccount<'account, 'multisig, MultisigSigner, Program>;
+    close_account::CloseAccount<'account, 'multisig, MultisigSigner, Program>;
 
 /// Freeze an Initialized account using the Mint's `freeze_authority` (if
 /// set).
@@ -148,7 +180,7 @@ pub type CloseAccount<'account, 'multisig, MultisigSigner> =
 ///   2. `[]` The mint's multisignature freeze authority.
 ///   3. `..+M` `[signer]` M signer accounts.
 pub type FreezeAccount<'account, 'multisig, MultisigSigner> =
-    crate::definitions::FreezeAccount<'account, 'multisig, MultisigSigner, Program>;
+    freeze_account::FreezeAccount<'account, 'multisig, MultisigSigner, Program>;
 
 /// Gets the required size of an account for the given mint as a
 /// little-endian `u64`.
@@ -159,7 +191,8 @@ pub type FreezeAccount<'account, 'multisig, MultisigSigner> =
 /// Accounts expected by this instruction:
 ///
 ///   0. `[]` The mint to calculate for.
-pub type GetAccountDataSize<'account> = crate::definitions::GetAccountDataSize<'account, Program>;
+pub type GetAccountDataSize<'account> =
+    get_account_data_size::GetAccountDataSize<'account, Program>;
 
 /// Initializes a new account to hold tokens.  If this account is associated
 /// with the native mint then the token balance of the initialized account
@@ -179,7 +212,7 @@ pub type GetAccountDataSize<'account> = crate::definitions::GetAccountDataSize<'
 ///   1. `[]` The mint this account will be associated with.
 ///   2. `[]` The new account's owner/multisignature.
 ///   3. `[]` Rent sysvar.
-pub type InitializeAccount<'account> = crate::definitions::InitializeAccount<'account, Program>;
+pub type InitializeAccount<'account> = initialize_account::InitializeAccount<'account, Program>;
 
 /// Like [`InitializeAccount`], but the owner pubkey is
 /// passed via instruction data rather than the accounts list. This
@@ -192,7 +225,7 @@ pub type InitializeAccount<'account> = crate::definitions::InitializeAccount<'ac
 ///   0. `[writable]`  The account to initialize.
 ///   1. `[]` The mint this account will be associated with.
 ///   2. `[]` Rent sysvar.
-pub type InitializeAccount2<'account> = crate::definitions::InitializeAccount2<'account, Program>;
+pub type InitializeAccount2<'account> = initialize_account2::InitializeAccount2<'account, Program>;
 
 /// Like [`InitializeAccount2`], but does not require the
 /// Rent sysvar to be provided
@@ -202,7 +235,7 @@ pub type InitializeAccount2<'account> = crate::definitions::InitializeAccount2<'
 ///   0. `[writable]`  The account to initialize.
 ///   1. `[]` The mint this account will be associated with.
 pub type InitializeAccount3<'account, 'address> =
-    crate::definitions::InitializeAccount3<'account, 'address, Program>;
+    initialize_account3::InitializeAccount3<'account, 'address, Program>;
 
 /// Initialize the Immutable Owner extension for the given token account
 ///
@@ -213,7 +246,7 @@ pub type InitializeAccount3<'account, 'address> =
 ///
 ///   0. `[writable]`  The account to initialize.
 pub type InitializeImmutableOwner<'account> =
-    crate::definitions::InitializeImmutableOwner<'account, Program>;
+    initialize_immutable_owner::InitializeImmutableOwner<'account, Program>;
 
 /// Initializes a new mint and optionally deposits all the newly minted
 /// tokens in an account.
@@ -229,7 +262,7 @@ pub type InitializeImmutableOwner<'account> =
 ///   0. `[writable]` The mint to initialize.
 ///   1. `[]` Rent sysvar.
 pub type InitializeMint<'account, 'address> =
-    crate::definitions::InitializeMint<'account, 'address, Program>;
+    initialize_mint::InitializeMint<'account, 'address, Program>;
 
 /// Like [`InitializeMint`], but does not require the Rent
 /// sysvar to be provided
@@ -238,7 +271,7 @@ pub type InitializeMint<'account, 'address> =
 ///
 ///   0. `[writable]` The mint to initialize.
 pub type InitializeMint2<'account, 'address> =
-    crate::definitions::InitializeMint2<'account, 'address, Program>;
+    initialize_mint2::InitializeMint2<'account, 'address, Program>;
 
 /// Initializes a multisignature account with N provided signers.
 ///
@@ -260,7 +293,7 @@ pub type InitializeMint2<'account, 'address> =
 ///   2. `..+N` `[signer]` The signer accounts, must equal to N where `1 <= N <=
 ///      11`.
 pub type InitializeMultisig<'account, 'multisig, MultisigSigner> =
-    crate::definitions::InitializeMultisig<'account, 'multisig, MultisigSigner, Program>;
+    initialize_multisig::InitializeMultisig<'account, 'multisig, MultisigSigner, Program>;
 
 /// Like [`InitializeMultisig`], but does not require the
 /// Rent sysvar to be provided
@@ -271,7 +304,7 @@ pub type InitializeMultisig<'account, 'multisig, MultisigSigner> =
 ///   1. `..+N` `[signer]` The signer accounts, must equal to N where `1 <= N <=
 ///      11`.
 pub type InitializeMultisig2<'account, 'multisig, MultisigSigner> =
-    crate::definitions::InitializeMultisig2<'account, 'multisig, MultisigSigner, Program>;
+    initialize_multisig2::InitializeMultisig2<'account, 'multisig, MultisigSigner, Program>;
 
 /// Mints new tokens to an account.  The native mint does not support
 /// minting.
@@ -289,7 +322,7 @@ pub type InitializeMultisig2<'account, 'multisig, MultisigSigner> =
 ///   2. `[]` The mint's multisignature mint-tokens authority.
 ///   3. `..+M` `[signer]` M signer accounts.
 pub type MintTo<'account, 'multisig, MultisigSigner> =
-    crate::definitions::MintTo<'account, 'multisig, MultisigSigner, Program>;
+    mint_to::MintTo<'account, 'multisig, MultisigSigner, Program>;
 
 /// Mints new tokens to an account.  The native mint does not support
 /// minting.
@@ -311,7 +344,7 @@ pub type MintTo<'account, 'multisig, MultisigSigner> =
 ///   2. `[]` The mint's multisignature mint-tokens authority.
 ///   3. `..+M` `[signer]` M signer accounts.
 pub type MintToChecked<'account, 'multisig, MultisigSigner> =
-    crate::definitions::MintToChecked<'account, 'multisig, MultisigSigner, Program>;
+    mint_to_checked::MintToChecked<'account, 'multisig, MultisigSigner, Program>;
 
 /// Revokes the delegate's authority.
 ///
@@ -326,7 +359,7 @@ pub type MintToChecked<'account, 'multisig, MultisigSigner> =
 ///   1. `[]` The source account's multisignature owner/delegate.
 ///   2. `..+M` `[signer]` M signer accounts.
 pub type Revoke<'account, 'multisig, MultisigSigner> =
-    crate::definitions::Revoke<'account, 'multisig, MultisigSigner, Program>;
+    revoke::Revoke<'account, 'multisig, MultisigSigner, Program>;
 
 /// Sets a new authority of a mint or account.
 ///
@@ -341,7 +374,7 @@ pub type Revoke<'account, 'multisig, MultisigSigner> =
 ///   1. `[]` The mint's or account's current multisignature authority.
 ///   2. `..+M` `[signer]` M signer accounts.
 pub type SetAuthority<'account, 'address, 'multisig, MultisigSigner> =
-    crate::definitions::SetAuthority<'account, 'address, 'multisig, MultisigSigner, Program>;
+    set_authority::SetAuthority<'account, 'address, 'multisig, MultisigSigner, Program>;
 
 /// Given a wrapped / native token account (a token account containing SOL)
 /// updates its amount field based on the account's underlying `lamports`.
@@ -359,7 +392,7 @@ pub type SetAuthority<'account, 'address, 'multisig, MultisigSigner> =
 ///   0. `[writable]`  The native token account to sync with its underlying
 ///      lamports.
 ///   1. `[]` Rent sysvar.
-pub type SyncNative<'account> = crate::definitions::SyncNative<'account, Program>;
+pub type SyncNative<'account> = sync_native::SyncNative<'account, Program>;
 
 /// Thaw a Frozen account using the Mint's `freeze_authority` (if set).
 ///
@@ -376,7 +409,7 @@ pub type SyncNative<'account> = crate::definitions::SyncNative<'account, Program
 ///   2. `[]` The mint's multisignature freeze authority.
 ///   3. `..+M` `[signer]` M signer accounts.
 pub type ThawAccount<'account, 'multisig, MultisigSigner> =
-    crate::definitions::ThawAccount<'account, 'multisig, MultisigSigner, Program>;
+    thaw_account::ThawAccount<'account, 'multisig, MultisigSigner, Program>;
 
 /// Transfers tokens from one account to another either directly or via a
 /// delegate.  If this account is associated with the native mint then equal
@@ -396,7 +429,7 @@ pub type ThawAccount<'account, 'multisig, MultisigSigner> =
 ///   2. `[]` The source account's multisignature owner/delegate.
 ///   3. `..+M` `[signer]` M signer accounts.
 pub type Transfer<'account, 'multisig, MultisigSigner> =
-    crate::definitions::Transfer<'account, 'multisig, MultisigSigner, Program>;
+    transfer::Transfer<'account, 'multisig, MultisigSigner, Program>;
 
 /// Transfers tokens from one account to another either directly or via a
 /// delegate.  If this account is associated with the native mint then equal
@@ -422,7 +455,7 @@ pub type Transfer<'account, 'multisig, MultisigSigner> =
 ///   3. `[]` The source account's multisignature owner/delegate.
 ///   4. `..+M` `[signer]` M signer accounts.
 pub type TransferChecked<'account, 'multisig, MultisigSigner> =
-    crate::definitions::TransferChecked<'account, 'multisig, MultisigSigner, Program>;
+    transfer_checked::TransferChecked<'account, 'multisig, MultisigSigner, Program>;
 
 /// Convert a `UiAmount` of tokens to a little-endian `u64` raw Amount,
 /// using the given mint. In this version of the program, the mint can
@@ -435,7 +468,7 @@ pub type TransferChecked<'account, 'multisig, MultisigSigner> =
 ///
 ///   0. `[]` The mint to calculate for.
 pub type UiAmountToAmount<'account, 'amount> =
-    crate::definitions::UiAmountToAmount<'account, 'amount, Program>;
+    ui_amount_to_amount::UiAmountToAmount<'account, 'amount, Program>;
 
 /// Transfer lamports from a native SOL account to a destination account.
 ///
@@ -454,7 +487,7 @@ pub type UiAmountToAmount<'account, 'amount> =
 ///   2. `[]` The source account's multisignature owner/delegate.
 ///   3. `..+M` `[signer]` M signer accounts.
 pub type UnwrapLamports<'account, 'multisig, MultisigSigner> =
-    crate::definitions::UnwrapLamports<'account, 'multisig, MultisigSigner, Program>;
+    unwrap_lamports::UnwrapLamports<'account, 'multisig, MultisigSigner, Program>;
 
 /// This instruction is to be used to rescue SOL sent to any `TokenProgram`
 /// owned account by sending them to any other account, leaving behind only
@@ -473,4 +506,58 @@ pub type UnwrapLamports<'account, 'multisig, MultisigSigner> =
 ///   2. `[]` The source account's multisignature owner/delegate.
 ///   3. `..+M` `[signer]` M signer accounts.
 pub type WithdrawExcessLamports<'account, 'multisig, MultisigSigner> =
-    crate::definitions::WithdrawExcessLamports<'account, 'multisig, MultisigSigner, Program>;
+    withdraw_excess_lamports::WithdrawExcessLamports<'account, 'multisig, MultisigSigner, Program>;
+
+#[cold]
+fn account_borrow_failed_error() -> ProgramError {
+    ProgramError::AccountBorrowFailed
+}
+
+#[cold]
+fn invalid_argument_error() -> ProgramError {
+    ProgramError::InvalidArgument
+}
+
+/// A helper function to write bytes from a source slice to a destination slice of `MaybeUninit<u8>`.
+#[inline(always)]
+fn write_bytes(destination: &mut [MaybeUninit<u8>], source: &[u8]) {
+    let len = destination.len().min(source.len());
+    // SAFETY:
+    // - Both pointers have alignment 1.
+    // - For valid (non-UB) references, the borrow checker guarantees no overlap.
+    // - `len` is bounded by both slice lengths.
+    unsafe {
+        core::ptr::copy_nonoverlapping(source.as_ptr(), destination.as_mut_ptr() as *mut u8, len);
+    }
+}
+
+/// A trait for instructions that can be used in a CPI context.
+pub trait CpiWriter {
+    /// Writes the `AccountView`s required by this instruction into the provided
+    /// slice.
+    ///
+    /// Returns the number of accounts written.
+    fn write_accounts<'source, 'cpi>(
+        &'source self,
+        accounts: &mut [MaybeUninit<CpiAccount<'cpi>>],
+    ) -> Result<usize, ProgramError>
+    where
+        Self: 'cpi;
+
+    /// Writes the `InstructionAccount`s required by this instruction into the
+    /// provided slice.
+    ///
+    /// Returns the number of accounts written.
+    fn write_instruction_accounts<'source, 'cpi>(
+        &'source self,
+        accounts: &mut [MaybeUninit<InstructionAccount<'cpi>>],
+    ) -> Result<usize, ProgramError>
+    where
+        Self: 'cpi;
+
+    /// Writes the instruction data for this instruction into the provided
+    /// slice.
+    ///
+    /// Returns the number of bytes written.
+    fn write_instruction_data(&self, data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError>;
+}
