+++ programs/token/src/instructions/revoke.rs
@@ -1,104 +1,120 @@
 use {
     crate::{
-        definitions::{
-            account_borrow_failed_error, invalid_argument_error, CpiWriter, TokenProgram,
+        instructions::{
+            account_borrow_failed_error, initialize_multisig::MAX_MULTISIG_SIGNERS,
+            invalid_argument_error, CpiWriter, UNINIT_BYTE, UNINIT_CPI_ACCOUNT,
+            UNINIT_INSTRUCTION_ACCOUNT,
         },
-        UNINIT_BYTE, UNINIT_CPI_ACCOUNT, UNINIT_INSTRUCTION_ACCOUNT,
+        TokenProgram,
     },
     core::{marker::PhantomData, mem::MaybeUninit, slice::from_raw_parts},
     solana_account_view::AccountView,
+    solana_address::Address,
     solana_instruction_view::{
-        cpi::{invoke_unchecked, CpiAccount},
+        cpi::{invoke_signed_unchecked, CpiAccount, Signer},
         InstructionAccount, InstructionView,
     },
     solana_program_error::{ProgramError, ProgramResult},
 };
 
-/// Maximum number of multisignature signers.
-pub const MAX_MULTISIG_SIGNERS: usize = 11;
-
 /// The instruction discriminator.
-const DISCRIMINATOR: u8 = 2;
+const DISCRIMINATOR: u8 = 5;
 
 /// Maximum number of accounts expected by this instruction.
 ///
-/// The required number of accounts will depend on the number of signer
-/// accounts.
+/// The required number of accounts will depend whether the
+/// source account has a single owner or a multisignature
+/// owner.
 const MAX_ACCOUNTS_LEN: usize = 2 + MAX_MULTISIG_SIGNERS;
 
 /// Instruction data length:
 ///   - discriminator (1 byte)
-///   - number of signers (1 byte)
-const DATA_LEN: usize = 2;
+const DATA_LEN: usize = 1;
 
-/// Initializes a multisignature account with N provided signers.
-///
-/// Multisignature accounts can used in place of any single owner/delegate
-/// accounts in any token instruction that require an owner/delegate to be
-/// present.  The variant field represents the number of signers (M)
-/// required to validate this multisignature account.
-///
-/// The [`super::InitializeMultisig`] instruction requires no
-/// signers and MUST be included within the same Transaction as the
-/// system program's `CreateAccount` instruction that creates the
-/// account being initialized. Otherwise another party can acquire
-/// ownership of the uninitialized account.
+/// Revokes the delegate's authority.
 ///
 /// Accounts expected by this instruction:
 ///
-///   0. `[writable]` The multisignature account to initialize.
-///   1. `[]` Rent sysvar.
-///   2. `..+N` `[signer]` The signer accounts, must equal to N where `1 <= N <=
-///      11`.
-pub struct InitializeMultisig<
-    'account,
-    'multisig,
-    MultisigSigner: AsRef<AccountView>,
-    Program: TokenProgram,
-> where
-    'account: 'multisig,
-{
-    /// The multisignature account to initialize.
-    pub multisig: &'account AccountView,
+///   * Single owner
+///   0. `[writable]` The source account.
+///   1. `[signer]` The source account's owner/delegate.
+///
+///   * Multisignature owner/delegate
+///   0. `[writable]` The source account.
+///   1. `[]` The source account's multisignature owner/delegate.
+///   2. `..+M` `[signer]` M signer accounts.
+pub struct Revoke<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram> {
+    /// The source account.
+    pub source: &'account AccountView,
 
-    /// Rent sysvar.
-    pub rent_sysvar: &'account AccountView,
+    /// The source account's owner/delegate.
+    pub authority: &'account AccountView,
 
-    /// The signer accounts.
+    /// Multisignature signers.
     pub multisig_signers: &'multisig [MultisigSigner],
-
-    /// The number of signers (M) required to validate this multisignature
-    /// account.
-    pub m: u8,
 
     _program: PhantomData<Program>,
 }
 
-impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram>
-    InitializeMultisig<'account, 'multisig, MultisigSigner, Program>
-where
-    'account: 'multisig,
-{
+impl<'account, Program: TokenProgram> Revoke<'account, '_, &'account AccountView, Program> {
+    /// The instruction discriminator.
+    pub const DISCRIMINATOR: u8 = DISCRIMINATOR;
+
+    /// Maximum number of accounts expected by this instruction.
+    pub const MAX_ACCOUNTS_LEN: usize = MAX_ACCOUNTS_LEN;
+
+    /// Instruction data length.
+    pub const DATA_LEN: usize = DATA_LEN;
+
+    /// Creates a new `Revoke` instruction with a single owner authority.
     #[inline(always)]
-    pub fn new(
-        multisig: &'account AccountView,
-        rent_sysvar: &'account AccountView,
+    pub fn new(source: &'account AccountView, authority: &'account AccountView) -> Self {
+        Self::with_multisig_signers(source, authority, &[])
+    }
+}
+
+impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram>
+    Revoke<'account, 'multisig, MultisigSigner, Program>
+{
+    /// Creates a new `Revoke` instruction with a
+    /// multisignature owner authority and signer accounts.
+    #[inline(always)]
+    pub fn with_multisig_signers(
+        source: &'account AccountView,
+        authority: &'account AccountView,
         multisig_signers: &'multisig [MultisigSigner],
-        m: u8,
     ) -> Self {
         Self {
-            multisig,
-            rent_sysvar,
+            source,
+            authority,
             multisig_signers,
-            m,
             _program: PhantomData,
         }
     }
 
     #[inline(always)]
     pub fn invoke(&self) -> ProgramResult {
+        self.invoke_with_program(&Program::ID)
+    }
+
+    #[inline(always)]
+    pub fn invoke_signed(&self, signers: &[Signer]) -> ProgramResult {
+        self.invoke_signed_with_program(signers, &Program::ID)
+    }
+
+    #[inline(always)]
+    pub fn invoke_with_program(&self, program: &Address) -> ProgramResult {
+        self.invoke_signed_with_program(&[], program)
+    }
+
+    #[inline(always)]
+    pub fn invoke_signed_with_program(
+        &self,
+        signers: &[Signer],
+        program: &Address,
+    ) -> ProgramResult {
         if self.multisig_signers.len() > MAX_MULTISIG_SIGNERS {
-            return Err(ProgramError::InvalidArgument);
+            Err(ProgramError::InvalidArgument)?;
         }
 
         let mut instruction_accounts = [UNINIT_INSTRUCTION_ACCOUNT; MAX_ACCOUNTS_LEN];
@@ -112,9 +128,9 @@ where
         let written_instruction_data = self.write_instruction_data(&mut instruction_data)?;
 
         unsafe {
-            invoke_unchecked(
+            invoke_signed_unchecked(
                 &InstructionView {
-                    program_id: &Program::id(),
+                    program_id: program,
                     accounts: from_raw_parts(
                         instruction_accounts.as_ptr() as _,
                         written_instruction_accounts,
@@ -122,6 +138,7 @@ where
                     data: from_raw_parts(instruction_data.as_ptr() as _, written_instruction_data),
                 },
                 from_raw_parts(accounts.as_ptr() as _, written_accounts),
+                signers,
             );
         }
 
@@ -130,7 +147,7 @@ where
 }
 
 impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
-    for InitializeMultisig<'_, '_, MultisigSigner, Program>
+    for Revoke<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn write_accounts<'cpi>(
@@ -140,12 +157,7 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
     where
         Self: 'cpi,
     {
-        write_accounts(
-            self.multisig,
-            self.rent_sysvar,
-            self.multisig_signers,
-            accounts,
-        )
+        write_accounts(self.source, self.authority, self.multisig_signers, accounts)
     }
 
     #[inline(always)]
@@ -156,57 +168,45 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
     where
         Self: 'cpi,
     {
-        write_instruction_accounts(
-            self.multisig,
-            self.rent_sysvar,
-            self.multisig_signers,
-            accounts,
-        )
+        write_instruction_accounts(self.source, self.authority, self.multisig_signers, accounts)
     }
 
     #[inline(always)]
     fn write_instruction_data(&self, data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
-        write_instruction_data(self.m, data)
+        write_instruction_data(data)
     }
 }
 
-impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch<Program>
-    for InitializeMultisig<'_, '_, MultisigSigner, Program>
+impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::batch::IntoBatch<Program>
+    for Revoke<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn into_batch<'account, 'state>(
         self,
-        batch: &mut super::Batch<'account, 'state, Program>,
+        batch: &mut super::batch::Batch<'account, 'state, Program>,
     ) -> ProgramResult
     where
         Self: 'account + 'state,
     {
         batch.push(
-            |accounts| {
-                write_accounts(
-                    self.multisig,
-                    self.rent_sysvar,
-                    self.multisig_signers,
-                    accounts,
-                )
-            },
+            |accounts| write_accounts(self.source, self.authority, self.multisig_signers, accounts),
             |accounts| {
                 write_instruction_accounts(
-                    self.multisig,
-                    self.rent_sysvar,
+                    self.source,
+                    self.authority,
                     self.multisig_signers,
                     accounts,
                 )
             },
-            |data| write_instruction_data(self.m, data),
+            write_instruction_data,
         )
     }
 }
 
 #[inline(always)]
 fn write_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
-    multisig: &'account AccountView,
-    rent_sysvar: &'account AccountView,
+    source: &'account AccountView,
+    authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<CpiAccount<'out>>],
 ) -> Result<usize, ProgramError>
@@ -220,13 +220,13 @@ where
         return Err(invalid_argument_error());
     }
 
-    if multisig.is_borrowed() {
+    if source.is_borrowed() {
         return Err(account_borrow_failed_error());
     }
 
-    CpiAccount::init_from_account_view(multisig, &mut accounts[0]);
+    CpiAccount::init_from_account_view(source, &mut accounts[0]);
 
-    CpiAccount::init_from_account_view(rent_sysvar, &mut accounts[1]);
+    CpiAccount::init_from_account_view(authority, &mut accounts[1]);
 
     for (account, signer) in accounts[2..expected_accounts]
         .iter_mut()
@@ -240,8 +240,8 @@ where
 
 #[inline(always)]
 fn write_instruction_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
-    multisig: &'account AccountView,
-    rent_sysvar: &'account AccountView,
+    source: &'account AccountView,
+    authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<InstructionAccount<'out>>],
 ) -> Result<usize, ProgramError>
@@ -255,29 +255,33 @@ where
         return Err(invalid_argument_error());
     }
 
-    accounts[0].write(InstructionAccount::writable(multisig.address()));
+    accounts[0].write(InstructionAccount::writable(source.address()));
 
-    accounts[1].write(InstructionAccount::readonly(rent_sysvar.address()));
+    accounts[1].write(InstructionAccount::new(
+        authority.address(),
+        false,
+        multisig_signers.is_empty(),
+    ));
 
     for (account, signer) in accounts[2..expected_accounts]
         .iter_mut()
         .zip(multisig_signers.iter())
     {
-        account.write(InstructionAccount::readonly(signer.as_ref().address()));
+        account.write(InstructionAccount::readonly_signer(
+            signer.as_ref().address(),
+        ));
     }
 
     Ok(expected_accounts)
 }
 
 #[inline(always)]
-fn write_instruction_data(m: u8, data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
+fn write_instruction_data(data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
     if data.len() < DATA_LEN {
         return Err(invalid_argument_error());
     }
 
     data[0].write(DISCRIMINATOR);
-
-    data[1].write(m);
 
     Ok(DATA_LEN)
 }
