+++ programs/token/src/instructions/set_authority.rs
@@ -1,13 +1,15 @@
 use {
     crate::{
-        definitions::{
-            account_borrow_failed_error, invalid_argument_error, CpiWriter, TokenProgram,
-            MAX_MULTISIG_SIGNERS,
+        instructions::{
+            account_borrow_failed_error, initialize_multisig::MAX_MULTISIG_SIGNERS,
+            invalid_argument_error, write_bytes, CpiWriter, UNINIT_BYTE, UNINIT_CPI_ACCOUNT,
+            UNINIT_INSTRUCTION_ACCOUNT,
         },
-        UNINIT_BYTE, UNINIT_CPI_ACCOUNT, UNINIT_INSTRUCTION_ACCOUNT,
+        TokenProgram,
     },
     core::{marker::PhantomData, mem::MaybeUninit, slice::from_raw_parts},
     solana_account_view::AccountView,
+    solana_address::Address,
     solana_instruction_view::{
         cpi::{invoke_signed_unchecked, CpiAccount, Signer},
         InstructionAccount, InstructionView,
@@ -16,94 +18,135 @@ use {
 };
 
 /// The instruction discriminator.
-const DISCRIMINATOR: u8 = 11;
+const DISCRIMINATOR: u8 = 6;
 
 /// Maximum number of accounts expected by this instruction.
 ///
 /// The required number of accounts will depend whether the
 /// source account has a single owner or a multisignature
 /// owner.
-const MAX_ACCOUNTS_LEN: usize = 3 + MAX_MULTISIG_SIGNERS;
+const MAX_ACCOUNTS_LEN: usize = 2 + MAX_MULTISIG_SIGNERS;
 
 /// Instruction data length:
 ///   - discriminator (1 byte)
-const DATA_LEN: usize = 1;
+///   - authority type (1 byte)
+///   - new authority (33 bytes, optional)
+const DATA_LEN: usize = 35;
 
-/// Thaw a Frozen account using the Mint's `freeze_authority` (if set).
+#[repr(u8)]
+#[derive(Clone, Copy)]
+pub enum AuthorityType {
+    MintTokens = 0,
+    FreezeAccount = 1,
+    AccountOwner = 2,
+    CloseAccount = 3,
+}
+
+/// Sets a new authority of a mint or account.
 ///
 /// Accounts expected by this instruction:
 ///
-///   * Single owner
-///   0. `[writable]` The account to thaw.
-///   1. `[]` The token mint.
-///   2. `[signer]` The mint freeze authority.
+///   * Single authority
+///   0. `[writable]` The mint or account to change the authority of.
+///   1. `[signer]` The current authority of the mint or account.
 ///
-///   * Multisignature owner
-///   0. `[writable]` The account to thaw.
-///   1. `[]` The token mint.
-///   2. `[]` The mint's multisignature freeze authority.
-///   3. `..+M` `[signer]` M signer accounts.
-pub struct ThawAccount<
+///   * Multisignature authority
+///   0. `[writable]` The mint or account to change the authority of.
+///   1. `[]` The mint's or account's current multisignature authority.
+///   2. `..+M` `[signer]` M signer accounts.
+pub struct SetAuthority<
     'account,
+    'address,
     'multisig,
     MultisigSigner: AsRef<AccountView>,
     Program: TokenProgram,
 > {
-    ///  The account to thaw.
+    /// The mint or account to change the authority of.
     pub account: &'account AccountView,
 
-    /// The token mint.
-    pub mint: &'account AccountView,
-
-    ///  The mint freeze authority.
-    pub freeze_authority: &'account AccountView,
+    ///  The current authority of the mint or account.
+    pub authority: &'account AccountView,
 
     /// Multisignature signers.
     pub multisig_signers: &'multisig [MultisigSigner],
 
+    /// The type of authority to update.
+    pub authority_type: AuthorityType,
+
+    /// The new authority.
+    pub new_authority: Option<&'address Address>,
+
     _program: PhantomData<Program>,
 }
 
-impl<'account, Program: TokenProgram> ThawAccount<'account, '_, &'account AccountView, Program> {
-    /// Creates a new `ThawAccount` instruction with a single freeze authority.
+impl<'account, 'address, Program: TokenProgram>
+    SetAuthority<'account, 'address, '_, &'account AccountView, Program>
+{
+    /// The instruction discriminator.
+    pub const DISCRIMINATOR: u8 = DISCRIMINATOR;
+
+    /// Maximum number of accounts expected by this instruction.
+    pub const MAX_ACCOUNTS_LEN: usize = MAX_ACCOUNTS_LEN;
+
+    /// Instruction data length.
+    pub const DATA_LEN: usize = DATA_LEN;
+
+    /// Creates a new `SetAuthority` instruction with a single authority.
     #[inline(always)]
     pub fn new(
         account: &'account AccountView,
-        mint: &'account AccountView,
-        freeze_authority: &'account AccountView,
+        authority: &'account AccountView,
+        authority_type: AuthorityType,
+        new_authority: Option<&'address Address>,
     ) -> Self {
-        Self::with_multisig_signers(account, mint, freeze_authority, &[])
+        Self::with_multisig_signers(account, authority, authority_type, new_authority, &[])
     }
 }
 
-impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram>
-    ThawAccount<'account, 'multisig, MultisigSigner, Program>
+impl<'account, 'address, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram>
+    SetAuthority<'account, 'address, 'multisig, MultisigSigner, Program>
 {
-    /// Creates a new `ThawAccount` instruction with a
-    /// multisignature freeze authority and signer accounts.
+    /// Creates a new `SetAuthority` instruction with a
+    /// multisignature authority and signer accounts.
     #[inline(always)]
     pub fn with_multisig_signers(
         account: &'account AccountView,
-        mint: &'account AccountView,
-        freeze_authority: &'account AccountView,
+        authority: &'account AccountView,
+        authority_type: AuthorityType,
+        new_authority: Option<&'address Address>,
         multisig_signers: &'multisig [MultisigSigner],
     ) -> Self {
         Self {
             account,
-            mint,
-            freeze_authority,
+            authority,
             multisig_signers,
+            authority_type,
+            new_authority,
             _program: PhantomData,
         }
     }
 
     #[inline(always)]
     pub fn invoke(&self) -> ProgramResult {
-        self.invoke_signed(&[])
+        self.invoke_with_program(&Program::ID)
     }
 
     #[inline(always)]
     pub fn invoke_signed(&self, signers: &[Signer]) -> ProgramResult {
+        self.invoke_signed_with_program(signers, &Program::ID)
+    }
+
+    #[inline(always)]
+    pub fn invoke_with_program(&self, program: &Address) -> ProgramResult {
+        self.invoke_signed_with_program(&[], program)
+    }
+
+    #[inline(always)]
+    pub fn invoke_signed_with_program(
+        &self,
+        signers: &[Signer],
+        program: &Address,
+    ) -> ProgramResult {
         if self.multisig_signers.len() > MAX_MULTISIG_SIGNERS {
             Err(ProgramError::InvalidArgument)?;
         }
@@ -121,7 +164,7 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
         unsafe {
             invoke_signed_unchecked(
                 &InstructionView {
-                    program_id: &Program::id(),
+                    program_id: program,
                     accounts: from_raw_parts(
                         instruction_accounts.as_ptr() as _,
                         written_instruction_accounts,
@@ -138,7 +181,7 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
 }
 
 impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
-    for ThawAccount<'_, '_, MultisigSigner, Program>
+    for SetAuthority<'_, '_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn write_accounts<'cpi>(
@@ -150,8 +193,7 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
     {
         write_accounts(
             self.account,
-            self.mint,
-            self.freeze_authority,
+            self.authority,
             self.multisig_signers,
             accounts,
         )
@@ -167,8 +209,7 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
     {
         write_instruction_accounts(
             self.account,
-            self.mint,
-            self.freeze_authority,
+            self.authority,
             self.multisig_signers,
             accounts,
         )
@@ -176,17 +217,17 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
 
     #[inline(always)]
     fn write_instruction_data(&self, data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
-        write_instruction_data(data)
+        write_instruction_data(self.authority_type, self.new_authority, data)
     }
 }
 
-impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch<Program>
-    for ThawAccount<'_, '_, MultisigSigner, Program>
+impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::batch::IntoBatch<Program>
+    for SetAuthority<'_, '_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn into_batch<'account, 'state>(
         self,
-        batch: &mut super::Batch<'account, 'state, Program>,
+        batch: &mut super::batch::Batch<'account, 'state, Program>,
     ) -> ProgramResult
     where
         Self: 'account + 'state,
@@ -195,8 +236,7 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch
             |accounts| {
                 write_accounts(
                     self.account,
-                    self.mint,
-                    self.freeze_authority,
+                    self.authority,
                     self.multisig_signers,
                     accounts,
                 )
@@ -204,13 +244,12 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch
             |accounts| {
                 write_instruction_accounts(
                     self.account,
-                    self.mint,
-                    self.freeze_authority,
+                    self.authority,
                     self.multisig_signers,
                     accounts,
                 )
             },
-            write_instruction_data,
+            |data| write_instruction_data(self.authority_type, self.new_authority, data),
         )
     }
 }
@@ -218,8 +257,7 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch
 #[inline(always)]
 fn write_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
     account: &'account AccountView,
-    mint: &'account AccountView,
-    freeze_authority: &'account AccountView,
+    authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<CpiAccount<'out>>],
 ) -> Result<usize, ProgramError>
@@ -227,7 +265,7 @@ where
     'account: 'out,
     'multisig: 'out,
 {
-    let expected_accounts = 3 + multisig_signers.len();
+    let expected_accounts = 2 + multisig_signers.len();
 
     if expected_accounts > accounts.len() {
         return Err(invalid_argument_error());
@@ -239,11 +277,9 @@ where
 
     CpiAccount::init_from_account_view(account, &mut accounts[0]);
 
-    CpiAccount::init_from_account_view(mint, &mut accounts[1]);
+    CpiAccount::init_from_account_view(authority, &mut accounts[1]);
 
-    CpiAccount::init_from_account_view(freeze_authority, &mut accounts[2]);
-
-    for (account, signer) in accounts[3..expected_accounts]
+    for (account, signer) in accounts[2..expected_accounts]
         .iter_mut()
         .zip(multisig_signers.iter())
     {
@@ -256,8 +292,7 @@ where
 #[inline(always)]
 fn write_instruction_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
     account: &'account AccountView,
-    mint: &'account AccountView,
-    freeze_authority: &'account AccountView,
+    authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<InstructionAccount<'out>>],
 ) -> Result<usize, ProgramError>
@@ -265,7 +300,7 @@ where
     'account: 'out,
     'multisig: 'out,
 {
-    let expected_accounts = 3 + multisig_signers.len();
+    let expected_accounts = 2 + multisig_signers.len();
 
     if expected_accounts > accounts.len() {
         return Err(invalid_argument_error());
@@ -273,15 +308,13 @@ where
 
     accounts[0].write(InstructionAccount::writable(account.address()));
 
-    accounts[1].write(InstructionAccount::readonly(mint.address()));
-
-    accounts[2].write(InstructionAccount::new(
-        freeze_authority.address(),
+    accounts[1].write(InstructionAccount::new(
+        authority.address(),
         false,
         multisig_signers.is_empty(),
     ));
 
-    for (account, signer) in accounts[3..expected_accounts]
+    for (account, signer) in accounts[2..expected_accounts]
         .iter_mut()
         .zip(multisig_signers.iter())
     {
@@ -294,12 +327,32 @@ where
 }
 
 #[inline(always)]
-fn write_instruction_data(data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
-    if data.len() < DATA_LEN {
+fn write_instruction_data(
+    authority_type: AuthorityType,
+    new_authority: Option<&Address>,
+    data: &mut [MaybeUninit<u8>],
+) -> Result<usize, ProgramError> {
+    if data.len() < 3 {
         return Err(invalid_argument_error());
     }
 
     data[0].write(DISCRIMINATOR);
 
-    Ok(DATA_LEN)
+    data[1].write(authority_type as u8);
+
+    if let Some(new_authority) = new_authority {
+        if data.len() < DATA_LEN {
+            return Err(invalid_argument_error());
+        }
+
+        data[2].write(1);
+
+        write_bytes(&mut data[3..DATA_LEN], new_authority.as_array());
+
+        Ok(DATA_LEN)
+    } else {
+        data[2].write(0);
+
+        Ok(3)
+    }
 }
