+++ programs/token/src/instructions/thaw_account.rs
@@ -1,13 +1,15 @@
 use {
     crate::{
-        definitions::{
-            account_borrow_failed_error, invalid_argument_error, CpiWriter, TokenProgram,
-            MAX_MULTISIG_SIGNERS,
+        instructions::{
+            account_borrow_failed_error, initialize_multisig::MAX_MULTISIG_SIGNERS,
+            invalid_argument_error, CpiWriter, UNINIT_BYTE, UNINIT_CPI_ACCOUNT,
+            UNINIT_INSTRUCTION_ACCOUNT,
         },
-        write_bytes, UNINIT_BYTE, UNINIT_CPI_ACCOUNT, UNINIT_INSTRUCTION_ACCOUNT,
+        TokenProgram,
     },
     core::{marker::PhantomData, mem::MaybeUninit, slice::from_raw_parts},
     solana_account_view::AccountView,
+    solana_address::Address,
     solana_instruction_view::{
         cpi::{invoke_signed_unchecked, CpiAccount, Signer},
         InstructionAccount, InstructionView,
@@ -16,7 +18,7 @@ use {
 };
 
 /// The instruction discriminator.
-const DISCRIMINATOR: u8 = 15;
+const DISCRIMINATOR: u8 = 11;
 
 /// Maximum number of accounts expected by this instruction.
 ///
@@ -27,104 +29,106 @@ const MAX_ACCOUNTS_LEN: usize = 3 + MAX_MULTISIG_SIGNERS;
 
 /// Instruction data length:
 ///   - discriminator (1 byte)
-///   - amount (8 bytes)
-///   - decimals (1 byte)
-const DATA_LEN: usize = 10;
+const DATA_LEN: usize = 1;
 
-/// Burns tokens by removing them from an account.
-/// [`super::BurnChecked`] does not support accounts
-/// associated with the native mint, use `CloseAccount` instead.
-///
-/// This instruction differs from Burn in that the decimals value is checked
-/// by the caller. This may be useful when creating transactions offline or
-/// within a hardware wallet.
+/// Thaw a Frozen account using the Mint's `freeze_authority` (if set).
 ///
 /// Accounts expected by this instruction:
 ///
-///   * Single owner/delegate
-///   0. `[writable]` The account to burn from.
-///   1. `[writable]` The token mint.
-///   2. `[signer]` The account's owner/delegate.
+///   * Single owner
+///   0. `[writable]` The account to thaw.
+///   1. `[]` The token mint.
+///   2. `[signer]` The mint freeze authority.
 ///
-///   * Multisignature owner/delegate
-///   0. `[writable]` The account to burn from.
-///   1. `[writable]` The token mint.
-///   2. `[]` The account's multisignature owner/delegate.
+///   * Multisignature owner
+///   0. `[writable]` The account to thaw.
+///   1. `[]` The token mint.
+///   2. `[]` The mint's multisignature freeze authority.
 ///   3. `..+M` `[signer]` M signer accounts.
-pub struct BurnChecked<
+pub struct ThawAccount<
     'account,
     'multisig,
     MultisigSigner: AsRef<AccountView>,
     Program: TokenProgram,
 > {
-    /// The account to burn from.
+    ///  The account to thaw.
     pub account: &'account AccountView,
 
     /// The token mint.
     pub mint: &'account AccountView,
 
-    /// The account's owner/delegate.
-    pub authority: &'account AccountView,
+    ///  The mint freeze authority.
+    pub freeze_authority: &'account AccountView,
 
     /// Multisignature signers.
     pub multisig_signers: &'multisig [MultisigSigner],
 
-    /// The amount of tokens to burn.
-    pub amount: u64,
-
-    /// Expected number of base 10 digits to the right of the decimal place.
-    pub decimals: u8,
-
     _program: PhantomData<Program>,
 }
 
-impl<'account, Program: TokenProgram> BurnChecked<'account, '_, &'account AccountView, Program> {
-    /// Creates a new `BurnChecked` instruction with a single
-    /// owner/delegate authority.
+impl<'account, Program: TokenProgram> ThawAccount<'account, '_, &'account AccountView, Program> {
+    /// The instruction discriminator.
+    pub const DISCRIMINATOR: u8 = DISCRIMINATOR;
+
+    /// Maximum number of accounts expected by this instruction.
+    pub const MAX_ACCOUNTS_LEN: usize = MAX_ACCOUNTS_LEN;
+
+    /// Instruction data length.
+    pub const DATA_LEN: usize = DATA_LEN;
+
+    /// Creates a new `ThawAccount` instruction with a single freeze authority.
     #[inline(always)]
     pub fn new(
         account: &'account AccountView,
         mint: &'account AccountView,
-        authority: &'account AccountView,
-        amount: u64,
-        decimals: u8,
+        freeze_authority: &'account AccountView,
     ) -> Self {
-        Self::with_multisig_signers(account, mint, authority, amount, decimals, &[])
+        Self::with_multisig_signers(account, mint, freeze_authority, &[])
     }
 }
 
 impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram>
-    BurnChecked<'account, 'multisig, MultisigSigner, Program>
+    ThawAccount<'account, 'multisig, MultisigSigner, Program>
 {
-    /// Creates a new `BurnChecked` instruction with a
-    /// multisignature owner/delegate authority and signer accounts.
+    /// Creates a new `ThawAccount` instruction with a
+    /// multisignature freeze authority and signer accounts.
     #[inline(always)]
     pub fn with_multisig_signers(
         account: &'account AccountView,
         mint: &'account AccountView,
-        authority: &'account AccountView,
-        amount: u64,
-        decimals: u8,
+        freeze_authority: &'account AccountView,
         multisig_signers: &'multisig [MultisigSigner],
     ) -> Self {
         Self {
             account,
             mint,
-            authority,
+            freeze_authority,
             multisig_signers,
-            amount,
-            decimals,
             _program: PhantomData,
         }
     }
 
     #[inline(always)]
     pub fn invoke(&self) -> ProgramResult {
-        self.invoke_signed(&[])
+        self.invoke_with_program(&Program::ID)
     }
 
     #[inline(always)]
     pub fn invoke_signed(&self, signers: &[Signer]) -> ProgramResult {
+        self.invoke_signed_with_program(signers, &Program::ID)
+    }
+
+    #[inline(always)]
+    pub fn invoke_with_program(&self, program: &Address) -> ProgramResult {
+        self.invoke_signed_with_program(&[], program)
+    }
+
+    #[inline(always)]
+    pub fn invoke_signed_with_program(
+        &self,
+        signers: &[Signer],
+        program: &Address,
+    ) -> ProgramResult {
         if self.multisig_signers.len() > MAX_MULTISIG_SIGNERS {
             Err(ProgramError::InvalidArgument)?;
         }
@@ -142,7 +146,7 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
         unsafe {
             invoke_signed_unchecked(
                 &InstructionView {
-                    program_id: &Program::id(),
+                    program_id: program,
                     accounts: from_raw_parts(
                         instruction_accounts.as_ptr() as _,
                         written_instruction_accounts,
@@ -159,7 +163,7 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
 }
 
 impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
-    for BurnChecked<'_, '_, MultisigSigner, Program>
+    for ThawAccount<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn write_accounts<'cpi>(
@@ -172,7 +176,7 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
         write_accounts(
             self.account,
             self.mint,
-            self.authority,
+            self.freeze_authority,
             self.multisig_signers,
             accounts,
         )
@@ -189,7 +193,7 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
         write_instruction_accounts(
             self.account,
             self.mint,
-            self.authority,
+            self.freeze_authority,
             self.multisig_signers,
             accounts,
         )
@@ -197,17 +201,17 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
 
     #[inline(always)]
     fn write_instruction_data(&self, data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
-        write_instruction_data(self.amount, self.decimals, data)
+        write_instruction_data(data)
     }
 }
 
-impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch<Program>
-    for BurnChecked<'_, '_, MultisigSigner, Program>
+impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::batch::IntoBatch<Program>
+    for ThawAccount<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn into_batch<'account, 'state>(
         self,
-        batch: &mut super::Batch<'account, 'state, Program>,
+        batch: &mut super::batch::Batch<'account, 'state, Program>,
     ) -> ProgramResult
     where
         Self: 'account + 'state,
@@ -217,7 +221,7 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch
                 write_accounts(
                     self.account,
                     self.mint,
-                    self.authority,
+                    self.freeze_authority,
                     self.multisig_signers,
                     accounts,
                 )
@@ -226,12 +230,12 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch
                 write_instruction_accounts(
                     self.account,
                     self.mint,
-                    self.authority,
+                    self.freeze_authority,
                     self.multisig_signers,
                     accounts,
                 )
             },
-            |data| write_instruction_data(self.amount, self.decimals, data),
+            write_instruction_data,
         )
     }
 }
@@ -240,7 +244,7 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch
 fn write_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
     account: &'account AccountView,
     mint: &'account AccountView,
-    authority: &'account AccountView,
+    freeze_authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<CpiAccount<'out>>],
 ) -> Result<usize, ProgramError>
@@ -254,7 +258,7 @@ where
         return Err(invalid_argument_error());
     }
 
-    if account.is_borrowed() | mint.is_borrowed() {
+    if account.is_borrowed() {
         return Err(account_borrow_failed_error());
     }
 
@@ -262,7 +266,7 @@ where
 
     CpiAccount::init_from_account_view(mint, &mut accounts[1]);
 
-    CpiAccount::init_from_account_view(authority, &mut accounts[2]);
+    CpiAccount::init_from_account_view(freeze_authority, &mut accounts[2]);
 
     for (account, signer) in accounts[3..expected_accounts]
         .iter_mut()
@@ -278,7 +282,7 @@ where
 fn write_instruction_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
     account: &'account AccountView,
     mint: &'account AccountView,
-    authority: &'account AccountView,
+    freeze_authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<InstructionAccount<'out>>],
 ) -> Result<usize, ProgramError>
@@ -294,10 +298,10 @@ where
 
     accounts[0].write(InstructionAccount::writable(account.address()));
 
-    accounts[1].write(InstructionAccount::writable(mint.address()));
+    accounts[1].write(InstructionAccount::readonly(mint.address()));
 
     accounts[2].write(InstructionAccount::new(
-        authority.address(),
+        freeze_authority.address(),
         false,
         multisig_signers.is_empty(),
     ));
@@ -315,20 +319,12 @@ where
 }
 
 #[inline(always)]
-fn write_instruction_data(
-    amount: u64,
-    decimals: u8,
-    data: &mut [MaybeUninit<u8>],
-) -> Result<usize, ProgramError> {
+fn write_instruction_data(data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
     if data.len() < DATA_LEN {
         return Err(invalid_argument_error());
     }
 
     data[0].write(DISCRIMINATOR);
-
-    write_bytes(&mut data[1..9], &amount.to_le_bytes());
-
-    data[9].write(decimals);
 
     Ok(DATA_LEN)
 }
