+++ programs/token/src/instructions/transfer.rs
@@ -1,13 +1,15 @@
 use {
     crate::{
-        definitions::{
-            account_borrow_failed_error, invalid_argument_error, CpiWriter, TokenProgram,
-            MAX_MULTISIG_SIGNERS,
+        instructions::{
+            account_borrow_failed_error, initialize_multisig::MAX_MULTISIG_SIGNERS,
+            invalid_argument_error, write_bytes, CpiWriter, UNINIT_BYTE, UNINIT_CPI_ACCOUNT,
+            UNINIT_INSTRUCTION_ACCOUNT,
         },
-        UNINIT_BYTE, UNINIT_CPI_ACCOUNT, UNINIT_INSTRUCTION_ACCOUNT,
+        TokenProgram,
     },
     core::{marker::PhantomData, mem::MaybeUninit, slice::from_raw_parts},
     solana_account_view::AccountView,
+    solana_address::Address,
     solana_instruction_view::{
         cpi::{invoke_signed_unchecked, CpiAccount, Signer},
         InstructionAccount, InstructionView,
@@ -16,34 +18,44 @@ use {
 };
 
 /// The instruction discriminator.
-const DISCRIMINATOR: u8 = 5;
+const DISCRIMINATOR: u8 = 3;
 
 /// Maximum number of accounts expected by this instruction.
 ///
 /// The required number of accounts will depend whether the
 /// source account has a single owner or a multisignature
 /// owner.
-const MAX_ACCOUNTS_LEN: usize = 2 + MAX_MULTISIG_SIGNERS;
+const MAX_ACCOUNTS_LEN: usize = 3 + MAX_MULTISIG_SIGNERS;
 
 /// Instruction data length:
 ///   - discriminator (1 byte)
-const DATA_LEN: usize = 1;
+///   - amount (8 bytes)
+const DATA_LEN: usize = 9;
 
-/// Revokes the delegate's authority.
+/// Transfers tokens from one account to another either directly or via a
+/// delegate.  If this account is associated with the native mint then equal
+/// amounts of SOL and Tokens will be transferred to the destination
+/// account.
 ///
 /// Accounts expected by this instruction:
 ///
-///   * Single owner
+///   * Single owner/delegate
 ///   0. `[writable]` The source account.
-///   1. `[signer]` The source account's owner/delegate.
+///   1. `[writable]` The destination account.
+///   2. `[signer]` The source account's owner/delegate.
 ///
 ///   * Multisignature owner/delegate
 ///   0. `[writable]` The source account.
-///   1. `[]` The source account's multisignature owner/delegate.
-///   2. `..+M` `[signer]` M signer accounts.
-pub struct Revoke<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram> {
+///   1. `[writable]` The destination account.
+///   2. `[]` The source account's multisignature owner/delegate.
+///   3. `..+M` `[signer]` M signer accounts.
+pub struct Transfer<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram>
+{
     /// The source account.
-    pub source: &'account AccountView,
+    pub from: &'account AccountView,
+
+    /// The destination account.
+    pub to: &'account AccountView,
 
     /// The source account's owner/delegate.
     pub authority: &'account AccountView,
@@ -51,43 +63,79 @@ pub struct Revoke<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Progr
     /// Multisignature signers.
     pub multisig_signers: &'multisig [MultisigSigner],
 
+    /// The amount of tokens to transfer.
+    pub amount: u64,
+
     _program: PhantomData<Program>,
 }
 
-impl<'account, Program: TokenProgram> Revoke<'account, '_, &'account AccountView, Program> {
-    /// Creates a new `Revoke` instruction with a single owner authority.
+impl<'account, Program: TokenProgram> Transfer<'account, '_, &'account AccountView, Program> {
+    /// The instruction discriminator.
+    pub const DISCRIMINATOR: u8 = DISCRIMINATOR;
+
+    /// Maximum number of accounts expected by this instruction.
+    pub const MAX_ACCOUNTS_LEN: usize = MAX_ACCOUNTS_LEN;
+
+    /// Instruction data length.
+    pub const DATA_LEN: usize = DATA_LEN;
+
+    /// Creates a new `Transfer` instruction with a single
+    /// owner/delegate authority.
     #[inline(always)]
-    pub fn new(source: &'account AccountView, authority: &'account AccountView) -> Self {
-        Self::with_multisig_signers(source, authority, &[])
+    pub fn new(
+        from: &'account AccountView,
+        to: &'account AccountView,
+        authority: &'account AccountView,
+        amount: u64,
+    ) -> Self {
+        Self::with_multisig_signers(from, to, authority, amount, &[])
     }
 }
 
 impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram>
-    Revoke<'account, 'multisig, MultisigSigner, Program>
+    Transfer<'account, 'multisig, MultisigSigner, Program>
 {
-    /// Creates a new `Revoke` instruction with a
-    /// multisignature owner authority and signer accounts.
+    /// Creates a new `Transfer` instruction with a
+    /// multisignature owner/delegate authority and signer accounts.
     #[inline(always)]
     pub fn with_multisig_signers(
-        source: &'account AccountView,
+        from: &'account AccountView,
+        to: &'account AccountView,
         authority: &'account AccountView,
+        amount: u64,
         multisig_signers: &'multisig [MultisigSigner],
     ) -> Self {
         Self {
-            source,
+            from,
+            to,
             authority,
             multisig_signers,
+            amount,
             _program: PhantomData,
         }
     }
 
     #[inline(always)]
     pub fn invoke(&self) -> ProgramResult {
-        self.invoke_signed(&[])
+        self.invoke_with_program(&Program::ID)
     }
 
     #[inline(always)]
     pub fn invoke_signed(&self, signers: &[Signer]) -> ProgramResult {
+        self.invoke_signed_with_program(signers, &Program::ID)
+    }
+
+    #[inline(always)]
+    pub fn invoke_with_program(&self, program: &Address) -> ProgramResult {
+        self.invoke_signed_with_program(&[], program)
+    }
+
+    #[inline(always)]
+    pub fn invoke_signed_with_program(
+        &self,
+        signers: &[Signer],
+        program: &Address,
+    ) -> ProgramResult {
         if self.multisig_signers.len() > MAX_MULTISIG_SIGNERS {
             Err(ProgramError::InvalidArgument)?;
         }
@@ -105,7 +153,7 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
         unsafe {
             invoke_signed_unchecked(
                 &InstructionView {
-                    program_id: &Program::id(),
+                    program_id: program,
                     accounts: from_raw_parts(
                         instruction_accounts.as_ptr() as _,
                         written_instruction_accounts,
@@ -122,7 +170,7 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
 }
 
 impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
-    for Revoke<'_, '_, MultisigSigner, Program>
+    for Transfer<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn write_accounts<'cpi>(
@@ -132,7 +180,13 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
     where
         Self: 'cpi,
     {
-        write_accounts(self.source, self.authority, self.multisig_signers, accounts)
+        write_accounts(
+            self.from,
+            self.to,
+            self.authority,
+            self.multisig_signers,
+            accounts,
+        )
     }
 
     #[inline(always)]
@@ -143,44 +197,60 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
     where
         Self: 'cpi,
     {
-        write_instruction_accounts(self.source, self.authority, self.multisig_signers, accounts)
+        write_instruction_accounts(
+            self.from,
+            self.to,
+            self.authority,
+            self.multisig_signers,
+            accounts,
+        )
     }
 
     #[inline(always)]
     fn write_instruction_data(&self, data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
-        write_instruction_data(data)
+        write_instruction_data(self.amount, data)
     }
 }
 
-impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch<Program>
-    for Revoke<'_, '_, MultisigSigner, Program>
+impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::batch::IntoBatch<Program>
+    for Transfer<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn into_batch<'account, 'state>(
         self,
-        batch: &mut super::Batch<'account, 'state, Program>,
+        batch: &mut super::batch::Batch<'account, 'state, Program>,
     ) -> ProgramResult
     where
         Self: 'account + 'state,
     {
         batch.push(
-            |accounts| write_accounts(self.source, self.authority, self.multisig_signers, accounts),
             |accounts| {
-                write_instruction_accounts(
-                    self.source,
+                write_accounts(
+                    self.from,
+                    self.to,
                     self.authority,
                     self.multisig_signers,
                     accounts,
                 )
             },
-            write_instruction_data,
+            |accounts| {
+                write_instruction_accounts(
+                    self.from,
+                    self.to,
+                    self.authority,
+                    self.multisig_signers,
+                    accounts,
+                )
+            },
+            |data| write_instruction_data(self.amount, data),
         )
     }
 }
 
 #[inline(always)]
 fn write_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
-    source: &'account AccountView,
+    from: &'account AccountView,
+    to: &'account AccountView,
     authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<CpiAccount<'out>>],
@@ -189,21 +259,23 @@ where
     'account: 'out,
     'multisig: 'out,
 {
-    let expected_accounts = 2 + multisig_signers.len();
+    let expected_accounts = 3 + multisig_signers.len();
 
     if expected_accounts > accounts.len() {
         return Err(invalid_argument_error());
     }
 
-    if source.is_borrowed() {
+    if from.is_borrowed() | to.is_borrowed() {
         return Err(account_borrow_failed_error());
     }
 
-    CpiAccount::init_from_account_view(source, &mut accounts[0]);
+    CpiAccount::init_from_account_view(from, &mut accounts[0]);
 
-    CpiAccount::init_from_account_view(authority, &mut accounts[1]);
+    CpiAccount::init_from_account_view(to, &mut accounts[1]);
 
-    for (account, signer) in accounts[2..expected_accounts]
+    CpiAccount::init_from_account_view(authority, &mut accounts[2]);
+
+    for (account, signer) in accounts[3..expected_accounts]
         .iter_mut()
         .zip(multisig_signers.iter())
     {
@@ -215,7 +287,8 @@ where
 
 #[inline(always)]
 fn write_instruction_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
-    source: &'account AccountView,
+    from: &'account AccountView,
+    to: &'account AccountView,
     authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<InstructionAccount<'out>>],
@@ -224,21 +297,23 @@ where
     'account: 'out,
     'multisig: 'out,
 {
-    let expected_accounts = 2 + multisig_signers.len();
+    let expected_accounts = 3 + multisig_signers.len();
 
     if expected_accounts > accounts.len() {
         return Err(invalid_argument_error());
     }
 
-    accounts[0].write(InstructionAccount::writable(source.address()));
+    accounts[0].write(InstructionAccount::writable(from.address()));
 
-    accounts[1].write(InstructionAccount::new(
+    accounts[1].write(InstructionAccount::writable(to.address()));
+
+    accounts[2].write(InstructionAccount::new(
         authority.address(),
         false,
         multisig_signers.is_empty(),
     ));
 
-    for (account, signer) in accounts[2..expected_accounts]
+    for (account, signer) in accounts[3..expected_accounts]
         .iter_mut()
         .zip(multisig_signers.iter())
     {
@@ -251,12 +326,17 @@ where
 }
 
 #[inline(always)]
-fn write_instruction_data(data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
+fn write_instruction_data(
+    amount: u64,
+    data: &mut [MaybeUninit<u8>],
+) -> Result<usize, ProgramError> {
     if data.len() < DATA_LEN {
         return Err(invalid_argument_error());
     }
 
     data[0].write(DISCRIMINATOR);
+
+    write_bytes(&mut data[1..DATA_LEN], &amount.to_le_bytes());
 
     Ok(DATA_LEN)
 }
