+++ programs/token/src/instructions/transfer_checked.rs
@@ -1,13 +1,15 @@
 use {
     crate::{
-        definitions::{
-            account_borrow_failed_error, invalid_argument_error, CpiWriter, TokenProgram,
-            MAX_MULTISIG_SIGNERS,
+        instructions::{
+            account_borrow_failed_error, initialize_multisig::MAX_MULTISIG_SIGNERS,
+            invalid_argument_error, write_bytes, CpiWriter, UNINIT_BYTE, UNINIT_CPI_ACCOUNT,
+            UNINIT_INSTRUCTION_ACCOUNT,
         },
-        write_bytes, UNINIT_BYTE, UNINIT_CPI_ACCOUNT, UNINIT_INSTRUCTION_ACCOUNT,
+        TokenProgram,
     },
     core::{marker::PhantomData, mem::MaybeUninit, slice::from_raw_parts},
     solana_account_view::AccountView,
+    solana_address::Address,
     solana_instruction_view::{
         cpi::{invoke_signed_unchecked, CpiAccount, Signer},
         InstructionAccount, InstructionView,
@@ -16,97 +18,149 @@ use {
 };
 
 /// The instruction discriminator.
-const DISCRIMINATOR: u8 = 7;
+const DISCRIMINATOR: u8 = 12;
 
 /// Maximum number of accounts expected by this instruction.
 ///
 /// The required number of accounts will depend whether the
 /// source account has a single owner or a multisignature
 /// owner.
-const MAX_ACCOUNTS_LEN: usize = 3 + MAX_MULTISIG_SIGNERS;
+const MAX_ACCOUNTS_LEN: usize = 4 + MAX_MULTISIG_SIGNERS;
 
 /// Instruction data length:
 ///   - discriminator (1 byte)
-///   - amount to mint (8 bytes)
-const DATA_LEN: usize = 9;
+///   - amount (8 bytes)
+///   - decimals (1 byte)
+const DATA_LEN: usize = 10;
 
-/// Mints new tokens to an account.  The native mint does not support
-/// minting.
+/// Transfers tokens from one account to another either directly or via a
+/// delegate.  If this account is associated with the native mint then equal
+/// amounts of SOL and Tokens will be transferred to the destination
+/// account.
+///
+/// This instruction differs from [`super::Transfer`] in that the token mint and
+/// decimals value is checked by the caller.  This may be useful when
+/// creating transactions offline or within a hardware wallet.
 ///
 /// Accounts expected by this instruction:
 ///
-///   * Single authority
-///   0. `[writable]` The mint.
-///   1. `[writable]` The account to mint tokens to.
-///   2. `[signer]` The mint's minting authority.
+///   * Single owner/delegate
+///   0. `[writable]` The source account.
+///   1. `[]` The token mint.
+///   2. `[writable]` The destination account.
+///   3. `[signer]` The source account's owner/delegate.
 ///
-///   * Multisignature authority
-///   0. `[writable]` The mint.
-///   1. `[writable]` The account to mint tokens to.
-///   2. `[]` The mint's multisignature mint-tokens authority.
-///   3. `..+M` `[signer]` M signer accounts.
-pub struct MintTo<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram> {
-    /// The mint.
+///   * Multisignature owner/delegate
+///   0. `[writable]` The source account.
+///   1. `[]` The token mint.
+///   2. `[writable]` The destination account.
+///   3. `[]` The source account's multisignature owner/delegate.
+///   4. `..+M` `[signer]` M signer accounts.
+pub struct TransferChecked<
+    'account,
+    'multisig,
+    MultisigSigner: AsRef<AccountView>,
+    Program: TokenProgram,
+> {
+    /// The source account.
+    pub from: &'account AccountView,
+
+    /// The token mint.
     pub mint: &'account AccountView,
 
-    /// The account to mint tokens to.
-    pub account: &'account AccountView,
+    ///  The destination account.
+    pub to: &'account AccountView,
 
-    /// The account to mint tokens to.
-    pub mint_authority: &'account AccountView,
+    /// The source account's owner/delegate.
+    pub authority: &'account AccountView,
 
     /// Multisignature signers.
     pub multisig_signers: &'multisig [MultisigSigner],
 
-    /// The amount of new tokens to mint.
+    /// The amount of tokens to transfer.
     pub amount: u64,
+
+    /// Expected number of base 10 digits to the right of the decimal place.
+    pub decimals: u8,
 
     _program: PhantomData<Program>,
 }
 
-impl<'account, Program: TokenProgram> MintTo<'account, '_, &'account AccountView, Program> {
-    /// Creates a new `MintTo` instruction with a single mint authority.
+impl<'account, Program: TokenProgram>
+    TransferChecked<'account, '_, &'account AccountView, Program>
+{
+    /// The instruction discriminator.
+    pub const DISCRIMINATOR: u8 = DISCRIMINATOR;
+
+    /// Maximum number of accounts expected by this instruction.
+    pub const MAX_ACCOUNTS_LEN: usize = MAX_ACCOUNTS_LEN;
+
+    /// Instruction data length.
+    pub const DATA_LEN: usize = DATA_LEN;
+
+    /// Creates a new `TransferChecked` instruction with a single
+    /// owner/delegate authority.
     #[inline(always)]
     pub fn new(
+        from: &'account AccountView,
         mint: &'account AccountView,
-        account: &'account AccountView,
-        mint_authority: &'account AccountView,
+        to: &'account AccountView,
+        authority: &'account AccountView,
         amount: u64,
+        decimals: u8,
     ) -> Self {
-        Self::with_multisig_signers(mint, account, mint_authority, amount, &[])
+        Self::with_multisig_signers(from, mint, to, authority, amount, decimals, &[])
     }
 }
 
 impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram>
-    MintTo<'account, 'multisig, MultisigSigner, Program>
+    TransferChecked<'account, 'multisig, MultisigSigner, Program>
 {
-    /// Creates a new `MintTo` instruction with a
-    /// multisignature mint authority and signer accounts.
+    /// Creates a new `TransferChecked` instruction with a
+    /// multisignature owner/delegate authority and signer accounts.
     #[inline(always)]
     pub fn with_multisig_signers(
+        from: &'account AccountView,
         mint: &'account AccountView,
-        account: &'account AccountView,
-        mint_authority: &'account AccountView,
+        to: &'account AccountView,
+        authority: &'account AccountView,
         amount: u64,
+        decimals: u8,
         multisig_signers: &'multisig [MultisigSigner],
     ) -> Self {
         Self {
+            from,
             mint,
-            account,
-            mint_authority,
+            to,
+            authority,
             multisig_signers,
             amount,
+            decimals,
             _program: PhantomData,
         }
     }
 
     #[inline(always)]
     pub fn invoke(&self) -> ProgramResult {
-        self.invoke_signed(&[])
+        self.invoke_with_program(&Program::ID)
     }
 
     #[inline(always)]
     pub fn invoke_signed(&self, signers: &[Signer]) -> ProgramResult {
+        self.invoke_signed_with_program(signers, &Program::ID)
+    }
+
+    #[inline(always)]
+    pub fn invoke_with_program(&self, program: &Address) -> ProgramResult {
+        self.invoke_signed_with_program(&[], program)
+    }
+
+    #[inline(always)]
+    pub fn invoke_signed_with_program(
+        &self,
+        signers: &[Signer],
+        program: &Address,
+    ) -> ProgramResult {
         if self.multisig_signers.len() > MAX_MULTISIG_SIGNERS {
             Err(ProgramError::InvalidArgument)?;
         }
@@ -124,7 +178,7 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
         unsafe {
             invoke_signed_unchecked(
                 &InstructionView {
-                    program_id: &Program::id(),
+                    program_id: program,
                     accounts: from_raw_parts(
                         instruction_accounts.as_ptr() as _,
                         written_instruction_accounts,
@@ -141,7 +195,7 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
 }
 
 impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
-    for MintTo<'_, '_, MultisigSigner, Program>
+    for TransferChecked<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn write_accounts<'cpi>(
@@ -152,9 +206,10 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
         Self: 'cpi,
     {
         write_accounts(
+            self.from,
             self.mint,
-            self.account,
-            self.mint_authority,
+            self.to,
+            self.authority,
             self.multisig_signers,
             accounts,
         )
@@ -169,9 +224,10 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
         Self: 'cpi,
     {
         write_instruction_accounts(
+            self.from,
             self.mint,
-            self.account,
-            self.mint_authority,
+            self.to,
+            self.authority,
             self.multisig_signers,
             accounts,
         )
@@ -179,17 +235,17 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
 
     #[inline(always)]
     fn write_instruction_data(&self, data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
-        write_instruction_data(self.amount, data)
+        write_instruction_data(self.amount, self.decimals, data)
     }
 }
 
-impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch<Program>
-    for MintTo<'_, '_, MultisigSigner, Program>
+impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::batch::IntoBatch<Program>
+    for TransferChecked<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn into_batch<'account, 'state>(
         self,
-        batch: &mut super::Batch<'account, 'state, Program>,
+        batch: &mut super::batch::Batch<'account, 'state, Program>,
     ) -> ProgramResult
     where
         Self: 'account + 'state,
@@ -197,32 +253,35 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch
         batch.push(
             |accounts| {
                 write_accounts(
+                    self.from,
                     self.mint,
-                    self.account,
-                    self.mint_authority,
+                    self.to,
+                    self.authority,
                     self.multisig_signers,
                     accounts,
                 )
             },
             |accounts| {
                 write_instruction_accounts(
+                    self.from,
                     self.mint,
-                    self.account,
-                    self.mint_authority,
+                    self.to,
+                    self.authority,
                     self.multisig_signers,
                     accounts,
                 )
             },
-            |data| write_instruction_data(self.amount, data),
+            |data| write_instruction_data(self.amount, self.decimals, data),
         )
     }
 }
 
 #[inline(always)]
 fn write_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
+    from: &'account AccountView,
     mint: &'account AccountView,
-    account: &'account AccountView,
-    mint_authority: &'account AccountView,
+    to: &'account AccountView,
+    authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<CpiAccount<'out>>],
 ) -> Result<usize, ProgramError>
@@ -230,23 +289,25 @@ where
     'account: 'out,
     'multisig: 'out,
 {
-    let expected_accounts = 3 + multisig_signers.len();
+    let expected_accounts = 4 + multisig_signers.len();
 
     if expected_accounts > accounts.len() {
         return Err(invalid_argument_error());
     }
 
-    if mint.is_borrowed() | account.is_borrowed() {
+    if from.is_borrowed() | to.is_borrowed() {
         return Err(account_borrow_failed_error());
     }
 
-    CpiAccount::init_from_account_view(mint, &mut accounts[0]);
+    CpiAccount::init_from_account_view(from, &mut accounts[0]);
 
-    CpiAccount::init_from_account_view(account, &mut accounts[1]);
+    CpiAccount::init_from_account_view(mint, &mut accounts[1]);
 
-    CpiAccount::init_from_account_view(mint_authority, &mut accounts[2]);
+    CpiAccount::init_from_account_view(to, &mut accounts[2]);
 
-    for (account, signer) in accounts[3..expected_accounts]
+    CpiAccount::init_from_account_view(authority, &mut accounts[3]);
+
+    for (account, signer) in accounts[4..expected_accounts]
         .iter_mut()
         .zip(multisig_signers.iter())
     {
@@ -258,9 +319,10 @@ where
 
 #[inline(always)]
 fn write_instruction_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
+    from: &'account AccountView,
     mint: &'account AccountView,
-    account: &'account AccountView,
-    mint_authority: &'account AccountView,
+    to: &'account AccountView,
+    authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<InstructionAccount<'out>>],
 ) -> Result<usize, ProgramError>
@@ -268,23 +330,25 @@ where
     'account: 'out,
     'multisig: 'out,
 {
-    let expected_accounts = 3 + multisig_signers.len();
+    let expected_accounts = 4 + multisig_signers.len();
 
     if expected_accounts > accounts.len() {
         return Err(invalid_argument_error());
     }
 
-    accounts[0].write(InstructionAccount::writable(mint.address()));
+    accounts[0].write(InstructionAccount::writable(from.address()));
 
-    accounts[1].write(InstructionAccount::writable(account.address()));
+    accounts[1].write(InstructionAccount::readonly(mint.address()));
 
-    accounts[2].write(InstructionAccount::new(
-        mint_authority.address(),
+    accounts[2].write(InstructionAccount::writable(to.address()));
+
+    accounts[3].write(InstructionAccount::new(
+        authority.address(),
         false,
         multisig_signers.is_empty(),
     ));
 
-    for (account, signer) in accounts[3..expected_accounts]
+    for (account, signer) in accounts[4..expected_accounts]
         .iter_mut()
         .zip(multisig_signers.iter())
     {
@@ -299,6 +363,7 @@ where
 #[inline(always)]
 fn write_instruction_data(
     amount: u64,
+    decimals: u8,
     data: &mut [MaybeUninit<u8>],
 ) -> Result<usize, ProgramError> {
     if data.len() < DATA_LEN {
@@ -307,7 +372,9 @@ fn write_instruction_data(
 
     data[0].write(DISCRIMINATOR);
 
-    write_bytes(&mut data[1..DATA_LEN], &amount.to_le_bytes());
+    write_bytes(&mut data[1..9], &amount.to_le_bytes());
+
+    data[9].write(decimals);
 
     Ok(DATA_LEN)
 }
