+++ programs/token/src/instructions/unwrap_lamports.rs
@@ -1,13 +1,15 @@
 use {
     crate::{
-        definitions::{
-            account_borrow_failed_error, invalid_argument_error, CpiWriter, TokenProgram,
-            MAX_MULTISIG_SIGNERS,
+        instructions::{
+            account_borrow_failed_error, initialize_multisig::MAX_MULTISIG_SIGNERS,
+            invalid_argument_error, write_bytes, CpiWriter, UNINIT_BYTE, UNINIT_CPI_ACCOUNT,
+            UNINIT_INSTRUCTION_ACCOUNT,
         },
-        write_bytes, UNINIT_BYTE, UNINIT_CPI_ACCOUNT, UNINIT_INSTRUCTION_ACCOUNT,
+        TokenProgram,
     },
     core::{marker::PhantomData, mem::MaybeUninit, slice::from_raw_parts},
     solana_account_view::AccountView,
+    solana_address::Address,
     solana_instruction_view::{
         cpi::{invoke_signed_unchecked, CpiAccount, Signer},
         InstructionAccount, InstructionView,
@@ -16,58 +18,62 @@ use {
 };
 
 /// The instruction discriminator.
-const DISCRIMINATOR: u8 = 12;
+const DISCRIMINATOR: u8 = 45;
 
 /// Maximum number of accounts expected by this instruction.
 ///
 /// The required number of accounts will depend whether the
 /// source account has a single owner or a multisignature
 /// owner.
-const MAX_ACCOUNTS_LEN: usize = 4 + MAX_MULTISIG_SIGNERS;
+const MAX_ACCOUNTS_LEN: usize = 3 + MAX_MULTISIG_SIGNERS;
 
 /// Instruction data length:
 ///   - discriminator (1 byte)
-///   - amount (8 bytes)
-///   - decimals (1 byte)
-const DATA_LEN: usize = 10;
+///   - amount (9 bytes, optional)
+const MAX_DATA_LEN: usize = 10;
 
-/// Transfers tokens from one account to another either directly or via a
-/// delegate.  If this account is associated with the native mint then equal
-/// amounts of SOL and Tokens will be transferred to the destination
-/// account.
+/// Enum specifying the amount of lamports to transfer
+/// from a native SOL account.
+#[repr(u8)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub enum Amount {
+    /// Transfer the entire amount of the source account.
+    All,
+
+    /// Transfer a specified amount.
+    ///
+    /// The value must be less than or equal to the amount
+    /// of the source account.
+    Some(u64),
+}
+
+/// Transfer lamports from a native SOL account to a destination account.
 ///
-/// This instruction differs from [`super::Transfer`] in that the token mint and
-/// decimals value is checked by the caller.  This may be useful when
-/// creating transactions offline or within a hardware wallet.
+/// This is useful to unwrap lamports from a wrapped SOL account.
 ///
 /// Accounts expected by this instruction:
 ///
 ///   * Single owner/delegate
 ///   0. `[writable]` The source account.
-///   1. `[]` The token mint.
-///   2. `[writable]` The destination account.
-///   3. `[signer]` The source account's owner/delegate.
+///   1. `[writable]` The destination account.
+///   2. `[signer]` The source account's owner/delegate.
 ///
 ///   * Multisignature owner/delegate
 ///   0. `[writable]` The source account.
-///   1. `[]` The token mint.
-///   2. `[writable]` The destination account.
-///   3. `[]` The source account's multisignature owner/delegate.
-///   4. `..+M` `[signer]` M signer accounts.
-pub struct TransferChecked<
+///   1. `[writable]` The destination account.
+///   2. `[]` The source account's multisignature owner/delegate.
+///   3. `..+M` `[signer]` M signer accounts.
+pub struct UnwrapLamports<
     'account,
     'multisig,
     MultisigSigner: AsRef<AccountView>,
     Program: TokenProgram,
 > {
     /// The source account.
-    pub from: &'account AccountView,
+    pub source: &'account AccountView,
 
-    /// The token mint.
-    pub mint: &'account AccountView,
-
-    ///  The destination account.
-    pub to: &'account AccountView,
+    /// The destination account.
+    pub destination: &'account AccountView,
 
     /// The source account's owner/delegate.
     pub authority: &'account AccountView,
@@ -75,67 +81,79 @@ pub struct TransferChecked<
     /// Multisignature signers.
     pub multisig_signers: &'multisig [MultisigSigner],
 
-    /// The amount of tokens to transfer.
-    pub amount: u64,
-
-    /// Expected number of base 10 digits to the right of the decimal place.
-    pub decimals: u8,
+    /// The amount of lamports to transfer.
+    pub amount: Amount,
 
     _program: PhantomData<Program>,
 }
 
-impl<'account, Program: TokenProgram>
-    TransferChecked<'account, '_, &'account AccountView, Program>
-{
-    /// Creates a new `TransferChecked` instruction with a single
-    /// owner/delegate authority.
+impl<'account, Program: TokenProgram> UnwrapLamports<'account, '_, &'account AccountView, Program> {
+    /// The instruction discriminator.
+    pub const DISCRIMINATOR: u8 = DISCRIMINATOR;
+
+    /// Maximum number of accounts expected by this instruction.
+    pub const MAX_ACCOUNTS_LEN: usize = MAX_ACCOUNTS_LEN;
+
+    /// Maximum instruction data length.
+    pub const MAX_DATA_LEN: usize = MAX_DATA_LEN;
+
+    /// Creates a new `UnwrapLamports` instruction with a single owner
+    /// authority.
     #[inline(always)]
     pub fn new(
-        from: &'account AccountView,
-        mint: &'account AccountView,
-        to: &'account AccountView,
+        source: &'account AccountView,
+        destination: &'account AccountView,
         authority: &'account AccountView,
-        amount: u64,
-        decimals: u8,
+        amount: Amount,
     ) -> Self {
-        Self::with_multisig_signers(from, mint, to, authority, amount, decimals, &[])
+        Self::with_multisig_signers(source, destination, authority, amount, &[])
     }
 }
 
 impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram>
-    TransferChecked<'account, 'multisig, MultisigSigner, Program>
+    UnwrapLamports<'account, 'multisig, MultisigSigner, Program>
 {
-    /// Creates a new `TransferChecked` instruction with a
-    /// multisignature owner/delegate authority and signer accounts.
+    /// Creates a new `UnwrapLamports` instruction with a
+    /// multisignature owner authority and signer accounts.
     #[inline(always)]
     pub fn with_multisig_signers(
-        from: &'account AccountView,
-        mint: &'account AccountView,
-        to: &'account AccountView,
+        source: &'account AccountView,
+        destination: &'account AccountView,
         authority: &'account AccountView,
-        amount: u64,
-        decimals: u8,
+        amount: Amount,
         multisig_signers: &'multisig [MultisigSigner],
     ) -> Self {
         Self {
-            from,
-            mint,
-            to,
+            source,
+            destination,
             authority,
             multisig_signers,
             amount,
-            decimals,
             _program: PhantomData,
         }
     }
 
     #[inline(always)]
     pub fn invoke(&self) -> ProgramResult {
-        self.invoke_signed(&[])
+        self.invoke_with_program(&Program::ID)
     }
 
     #[inline(always)]
     pub fn invoke_signed(&self, signers: &[Signer]) -> ProgramResult {
+        self.invoke_signed_with_program(signers, &Program::ID)
+    }
+
+    #[inline(always)]
+    pub fn invoke_with_program(&self, program: &Address) -> ProgramResult {
+        self.invoke_signed_with_program(&[], program)
+    }
+
+    #[inline(always)]
+    pub fn invoke_signed_with_program(
+        &self,
+        signers: &[Signer],
+        program: &Address,
+    ) -> ProgramResult {
         if self.multisig_signers.len() > MAX_MULTISIG_SIGNERS {
             Err(ProgramError::InvalidArgument)?;
         }
@@ -147,13 +165,13 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
         let mut accounts = [UNINIT_CPI_ACCOUNT; MAX_ACCOUNTS_LEN];
         let written_accounts = self.write_accounts(&mut accounts)?;
 
-        let mut instruction_data = [UNINIT_BYTE; DATA_LEN];
+        let mut instruction_data = [UNINIT_BYTE; MAX_DATA_LEN];
         let written_instruction_data = self.write_instruction_data(&mut instruction_data)?;
 
         unsafe {
             invoke_signed_unchecked(
                 &InstructionView {
-                    program_id: &Program::id(),
+                    program_id: program,
                     accounts: from_raw_parts(
                         instruction_accounts.as_ptr() as _,
                         written_instruction_accounts,
@@ -170,7 +188,7 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
 }
 
 impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
-    for TransferChecked<'_, '_, MultisigSigner, Program>
+    for UnwrapLamports<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn write_accounts<'cpi>(
@@ -181,9 +199,8 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
         Self: 'cpi,
     {
         write_accounts(
-            self.from,
-            self.mint,
-            self.to,
+            self.source,
+            self.destination,
             self.authority,
             self.multisig_signers,
             accounts,
@@ -199,9 +216,8 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
         Self: 'cpi,
     {
         write_instruction_accounts(
-            self.from,
-            self.mint,
-            self.to,
+            self.source,
+            self.destination,
             self.authority,
             self.multisig_signers,
             accounts,
@@ -210,17 +226,17 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
 
     #[inline(always)]
     fn write_instruction_data(&self, data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
-        write_instruction_data(self.amount, self.decimals, data)
+        write_instruction_data(self.amount, data)
     }
 }
 
-impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch<Program>
-    for TransferChecked<'_, '_, MultisigSigner, Program>
+impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::batch::IntoBatch<Program>
+    for UnwrapLamports<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn into_batch<'account, 'state>(
         self,
-        batch: &mut super::Batch<'account, 'state, Program>,
+        batch: &mut super::batch::Batch<'account, 'state, Program>,
     ) -> ProgramResult
     where
         Self: 'account + 'state,
@@ -228,9 +244,8 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch
         batch.push(
             |accounts| {
                 write_accounts(
-                    self.from,
-                    self.mint,
-                    self.to,
+                    self.source,
+                    self.destination,
                     self.authority,
                     self.multisig_signers,
                     accounts,
@@ -238,24 +253,22 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch
             },
             |accounts| {
                 write_instruction_accounts(
-                    self.from,
-                    self.mint,
-                    self.to,
+                    self.source,
+                    self.destination,
                     self.authority,
                     self.multisig_signers,
                     accounts,
                 )
             },
-            |data| write_instruction_data(self.amount, self.decimals, data),
+            |data| write_instruction_data(self.amount, data),
         )
     }
 }
 
 #[inline(always)]
 fn write_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
-    from: &'account AccountView,
-    mint: &'account AccountView,
-    to: &'account AccountView,
+    source: &'account AccountView,
+    destination: &'account AccountView,
     authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<CpiAccount<'out>>],
@@ -264,25 +277,23 @@ where
     'account: 'out,
     'multisig: 'out,
 {
-    let expected_accounts = 4 + multisig_signers.len();
+    let expected_accounts = 3 + multisig_signers.len();
 
     if expected_accounts > accounts.len() {
         return Err(invalid_argument_error());
     }
 
-    if from.is_borrowed() | to.is_borrowed() {
+    if source.is_borrowed() | destination.is_borrowed() {
         return Err(account_borrow_failed_error());
     }
 
-    CpiAccount::init_from_account_view(from, &mut accounts[0]);
+    CpiAccount::init_from_account_view(source, &mut accounts[0]);
 
-    CpiAccount::init_from_account_view(mint, &mut accounts[1]);
+    CpiAccount::init_from_account_view(destination, &mut accounts[1]);
 
-    CpiAccount::init_from_account_view(to, &mut accounts[2]);
+    CpiAccount::init_from_account_view(authority, &mut accounts[2]);
 
-    CpiAccount::init_from_account_view(authority, &mut accounts[3]);
-
-    for (account, signer) in accounts[4..expected_accounts]
+    for (account, signer) in accounts[3..expected_accounts]
         .iter_mut()
         .zip(multisig_signers.iter())
     {
@@ -294,9 +305,8 @@ where
 
 #[inline(always)]
 fn write_instruction_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
-    from: &'account AccountView,
-    mint: &'account AccountView,
-    to: &'account AccountView,
+    source: &'account AccountView,
+    destination: &'account AccountView,
     authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<InstructionAccount<'out>>],
@@ -305,25 +315,23 @@ where
     'account: 'out,
     'multisig: 'out,
 {
-    let expected_accounts = 4 + multisig_signers.len();
+    let expected_accounts = 3 + multisig_signers.len();
 
     if expected_accounts > accounts.len() {
         return Err(invalid_argument_error());
     }
 
-    accounts[0].write(InstructionAccount::writable(from.address()));
+    accounts[0].write(InstructionAccount::writable(source.address()));
 
-    accounts[1].write(InstructionAccount::readonly(mint.address()));
+    accounts[1].write(InstructionAccount::writable(destination.address()));
 
-    accounts[2].write(InstructionAccount::writable(to.address()));
-
-    accounts[3].write(InstructionAccount::new(
+    accounts[2].write(InstructionAccount::new(
         authority.address(),
         false,
         multisig_signers.is_empty(),
     ));
 
-    for (account, signer) in accounts[4..expected_accounts]
+    for (account, signer) in accounts[3..expected_accounts]
         .iter_mut()
         .zip(multisig_signers.iter())
     {
@@ -337,19 +345,24 @@ where
 
 #[inline(always)]
 fn write_instruction_data(
-    amount: u64,
-    decimals: u8,
+    amount: Amount,
     data: &mut [MaybeUninit<u8>],
 ) -> Result<usize, ProgramError> {
-    if data.len() < DATA_LEN {
+    if data.len() < MAX_DATA_LEN {
         return Err(invalid_argument_error());
     }
 
     data[0].write(DISCRIMINATOR);
 
-    write_bytes(&mut data[1..9], &amount.to_le_bytes());
+    if let Amount::Some(amount) = amount {
+        data[1].write(1);
 
-    data[9].write(decimals);
+        write_bytes(&mut data[2..MAX_DATA_LEN], &amount.to_le_bytes());
 
-    Ok(DATA_LEN)
+        Ok(MAX_DATA_LEN)
+    } else {
+        data[1].write(0);
+
+        Ok(2)
+    }
 }
