+++ programs/token/src/instructions/withdraw_excess_lamports.rs
@@ -1,13 +1,15 @@
 use {
     crate::{
-        definitions::{
-            account_borrow_failed_error, invalid_argument_error, CpiWriter, TokenProgram,
-            MAX_MULTISIG_SIGNERS,
+        instructions::{
+            account_borrow_failed_error, initialize_multisig::MAX_MULTISIG_SIGNERS,
+            invalid_argument_error, CpiWriter, UNINIT_BYTE, UNINIT_CPI_ACCOUNT,
+            UNINIT_INSTRUCTION_ACCOUNT,
         },
-        write_bytes, UNINIT_BYTE, UNINIT_CPI_ACCOUNT, UNINIT_INSTRUCTION_ACCOUNT,
+        TokenProgram,
     },
     core::{marker::PhantomData, mem::MaybeUninit, slice::from_raw_parts},
     solana_account_view::AccountView,
+    solana_address::Address,
     solana_instruction_view::{
         cpi::{invoke_signed_unchecked, CpiAccount, Signer},
         InstructionAccount, InstructionView,
@@ -16,7 +18,7 @@ use {
 };
 
 /// The instruction discriminator.
-const DISCRIMINATOR: u8 = 14;
+const DISCRIMINATOR: u8 = 38;
 
 /// Maximum number of accounts expected by this instruction.
 ///
@@ -27,103 +29,111 @@ const MAX_ACCOUNTS_LEN: usize = 3 + MAX_MULTISIG_SIGNERS;
 
 /// Instruction data length:
 ///   - discriminator (1 byte)
-///   - amount to mint (8 bytes)
-///   - decimals (1 byte)
-const DATA_LEN: usize = 10;
+const DATA_LEN: usize = 1;
 
-/// Mints new tokens to an account.  The native mint does not support
-/// minting.
-///
-/// This instruction differs from [`super::MintTo`] in that the
-/// decimals value is checked by the caller.  This may be useful when
-/// creating transactions offline or within a hardware wallet.
+/// This instruction is to be used to rescue SOL sent to any `TokenProgram`
+/// owned account by sending them to any other account, leaving behind only
+/// lamports for rent exemption.
 ///
 /// Accounts expected by this instruction:
 ///
-///   * Single authority
-///   0. `[writable]` The mint.
-///   1. `[writable]` The account to mint tokens to.
-///   2. `[signer]` The mint's minting authority.
+///   * Single owner/delegate
+///   0. `[writable]` The source account.
+///   1. `[writable]` The destination account.
+///   2. `[signer]` The source account's owner/delegate.
 ///
-///   * Multisignature authority
-///   0. `[writable]` The mint.
-///   1. `[writable]` The account to mint tokens to.
-///   2. `[]` The mint's multisignature mint-tokens authority.
+///   * Multisignature owner/delegate
+///   0. `[writable]` The source account.
+///   1. `[writable]` The destination account.
+///   2. `[]` The source account's multisignature owner/delegate.
 ///   3. `..+M` `[signer]` M signer accounts.
-pub struct MintToChecked<
+pub struct WithdrawExcessLamports<
     'account,
     'multisig,
     MultisigSigner: AsRef<AccountView>,
     Program: TokenProgram,
 > {
-    /// The mint.
-    pub mint: &'account AccountView,
+    /// Source Account owned by the token program.
+    pub source: &'account AccountView,
 
-    /// The account to mint tokens to.
-    pub account: &'account AccountView,
+    /// Destination account.
+    pub destination: &'account AccountView,
 
-    /// The mint's minting authority.
-    pub mint_authority: &'account AccountView,
+    /// Authority.
+    pub authority: &'account AccountView,
 
     /// Multisignature signers.
     pub multisig_signers: &'multisig [MultisigSigner],
 
-    /// The amount of new tokens to mint.
-    pub amount: u64,
-
-    /// Expected number of base 10 digits to the right of the decimal
-    ///     place.
-    pub decimals: u8,
-
     _program: PhantomData<Program>,
 }
 
-impl<'account, Program: TokenProgram> MintToChecked<'account, '_, &'account AccountView, Program> {
-    /// Creates a new `MintToChecked` instruction with a single mint authority.
+impl<'account, Program: TokenProgram>
+    WithdrawExcessLamports<'account, '_, &'account AccountView, Program>
+{
+    /// The instruction discriminator.
+    pub const DISCRIMINATOR: u8 = DISCRIMINATOR;
+
+    /// Maximum number of accounts expected by this instruction.
+    pub const MAX_ACCOUNTS_LEN: usize = MAX_ACCOUNTS_LEN;
+
+    /// Instruction data length.
+    pub const DATA_LEN: usize = DATA_LEN;
+
+    /// Creates a new `WithdrawExcessLamports` instruction with a single owner
+    /// authority.
     #[inline(always)]
     pub fn new(
-        mint: &'account AccountView,
-        account: &'account AccountView,
-        mint_authority: &'account AccountView,
-        amount: u64,
-        decimals: u8,
+        source: &'account AccountView,
+        destination: &'account AccountView,
+        authority: &'account AccountView,
     ) -> Self {
-        Self::with_multisig_signers(mint, account, mint_authority, amount, decimals, &[])
+        Self::with_multisig_signers(source, destination, authority, &[])
     }
 }
 
 impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProgram>
-    MintToChecked<'account, 'multisig, MultisigSigner, Program>
+    WithdrawExcessLamports<'account, 'multisig, MultisigSigner, Program>
 {
-    /// Creates a new `MintToChecked` instruction with a
-    /// multisignature mint authority and signer accounts.
+    /// Creates a new `WithdrawExcessLamports` instruction with a
+    /// multisignature owner authority and signer accounts.
     #[inline(always)]
     pub fn with_multisig_signers(
-        mint: &'account AccountView,
-        account: &'account AccountView,
-        mint_authority: &'account AccountView,
-        amount: u64,
-        decimals: u8,
+        source: &'account AccountView,
+        destination: &'account AccountView,
+        authority: &'account AccountView,
         multisig_signers: &'multisig [MultisigSigner],
     ) -> Self {
         Self {
-            mint,
-            account,
-            mint_authority,
+            source,
+            destination,
+            authority,
             multisig_signers,
-            amount,
-            decimals,
             _program: PhantomData,
         }
     }
 
     #[inline(always)]
     pub fn invoke(&self) -> ProgramResult {
-        self.invoke_signed(&[])
+        self.invoke_with_program(&Program::ID)
     }
 
     #[inline(always)]
     pub fn invoke_signed(&self, signers: &[Signer]) -> ProgramResult {
+        self.invoke_signed_with_program(signers, &Program::ID)
+    }
+
+    #[inline(always)]
+    pub fn invoke_with_program(&self, program: &Address) -> ProgramResult {
+        self.invoke_signed_with_program(&[], program)
+    }
+
+    #[inline(always)]
+    pub fn invoke_signed_with_program(
+        &self,
+        signers: &[Signer],
+        program: &Address,
+    ) -> ProgramResult {
         if self.multisig_signers.len() > MAX_MULTISIG_SIGNERS {
             Err(ProgramError::InvalidArgument)?;
         }
@@ -141,7 +151,7 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
         unsafe {
             invoke_signed_unchecked(
                 &InstructionView {
-                    program_id: &Program::id(),
+                    program_id: program,
                     accounts: from_raw_parts(
                         instruction_accounts.as_ptr() as _,
                         written_instruction_accounts,
@@ -158,7 +168,7 @@ impl<'account, 'multisig, MultisigSigner: AsRef<AccountView>, Program: TokenProg
 }
 
 impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
-    for MintToChecked<'_, '_, MultisigSigner, Program>
+    for WithdrawExcessLamports<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn write_accounts<'cpi>(
@@ -169,9 +179,9 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
         Self: 'cpi,
     {
         write_accounts(
-            self.mint,
-            self.account,
-            self.mint_authority,
+            self.source,
+            self.destination,
+            self.authority,
             self.multisig_signers,
             accounts,
         )
@@ -186,9 +196,9 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
         Self: 'cpi,
     {
         write_instruction_accounts(
-            self.mint,
-            self.account,
-            self.mint_authority,
+            self.source,
+            self.destination,
+            self.authority,
             self.multisig_signers,
             accounts,
         )
@@ -196,17 +206,17 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> CpiWriter
 
     #[inline(always)]
     fn write_instruction_data(&self, data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
-        write_instruction_data(self.amount, self.decimals, data)
+        write_instruction_data(data)
     }
 }
 
-impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch<Program>
-    for MintToChecked<'_, '_, MultisigSigner, Program>
+impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::batch::IntoBatch<Program>
+    for WithdrawExcessLamports<'_, '_, MultisigSigner, Program>
 {
     #[inline(always)]
     fn into_batch<'account, 'state>(
         self,
-        batch: &mut super::Batch<'account, 'state, Program>,
+        batch: &mut super::batch::Batch<'account, 'state, Program>,
     ) -> ProgramResult
     where
         Self: 'account + 'state,
@@ -214,32 +224,32 @@ impl<MultisigSigner: AsRef<AccountView>, Program: TokenProgram> super::IntoBatch
         batch.push(
             |accounts| {
                 write_accounts(
-                    self.mint,
-                    self.account,
-                    self.mint_authority,
+                    self.source,
+                    self.destination,
+                    self.authority,
                     self.multisig_signers,
                     accounts,
                 )
             },
             |accounts| {
                 write_instruction_accounts(
-                    self.mint,
-                    self.account,
-                    self.mint_authority,
+                    self.source,
+                    self.destination,
+                    self.authority,
                     self.multisig_signers,
                     accounts,
                 )
             },
-            |data| write_instruction_data(self.amount, self.decimals, data),
+            write_instruction_data,
         )
     }
 }
 
 #[inline(always)]
 fn write_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
-    mint: &'account AccountView,
-    account: &'account AccountView,
-    mint_authority: &'account AccountView,
+    source: &'account AccountView,
+    destination: &'account AccountView,
+    authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<CpiAccount<'out>>],
 ) -> Result<usize, ProgramError>
@@ -253,15 +263,15 @@ where
         return Err(invalid_argument_error());
     }
 
-    if mint.is_borrowed() | account.is_borrowed() {
+    if source.is_borrowed() | destination.is_borrowed() {
         return Err(account_borrow_failed_error());
     }
 
-    CpiAccount::init_from_account_view(mint, &mut accounts[0]);
+    CpiAccount::init_from_account_view(source, &mut accounts[0]);
 
-    CpiAccount::init_from_account_view(account, &mut accounts[1]);
+    CpiAccount::init_from_account_view(destination, &mut accounts[1]);
 
-    CpiAccount::init_from_account_view(mint_authority, &mut accounts[2]);
+    CpiAccount::init_from_account_view(authority, &mut accounts[2]);
 
     for (account, signer) in accounts[3..expected_accounts]
         .iter_mut()
@@ -275,9 +285,9 @@ where
 
 #[inline(always)]
 fn write_instruction_accounts<'account, 'multisig, 'out, MultisigSigner: AsRef<AccountView>>(
-    mint: &'account AccountView,
-    account: &'account AccountView,
-    mint_authority: &'account AccountView,
+    source: &'account AccountView,
+    destination: &'account AccountView,
+    authority: &'account AccountView,
     multisig_signers: &'multisig [MultisigSigner],
     accounts: &mut [MaybeUninit<InstructionAccount<'out>>],
 ) -> Result<usize, ProgramError>
@@ -291,12 +301,12 @@ where
         return Err(invalid_argument_error());
     }
 
-    accounts[0].write(InstructionAccount::writable(mint.address()));
+    accounts[0].write(InstructionAccount::writable(source.address()));
 
-    accounts[1].write(InstructionAccount::writable(account.address()));
+    accounts[1].write(InstructionAccount::writable(destination.address()));
 
     accounts[2].write(InstructionAccount::new(
-        mint_authority.address(),
+        authority.address(),
         false,
         multisig_signers.is_empty(),
     ));
@@ -314,20 +324,12 @@ where
 }
 
 #[inline(always)]
-fn write_instruction_data(
-    amount: u64,
-    decimals: u8,
-    data: &mut [MaybeUninit<u8>],
-) -> Result<usize, ProgramError> {
+fn write_instruction_data(data: &mut [MaybeUninit<u8>]) -> Result<usize, ProgramError> {
     if data.len() < DATA_LEN {
         return Err(invalid_argument_error());
     }
 
     data[0].write(DISCRIMINATOR);
-
-    write_bytes(&mut data[1..9], &amount.to_le_bytes());
-
-    data[9].write(decimals);
 
     Ok(DATA_LEN)
 }
